Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4458005 0.4228078 0.4607649 0.4387327 0.4373671 0.4228078 0.4567077
#>   [8] 0.4369539 0.4821932 0.4611302 0.4158415 0.4657229 0.4201637 0.4235727
#>  [15] 0.4219502 0.4082627 0.4758090 0.4168280 0.4758696 0.4614468 0.4321517
#>  [22] 0.4980526 0.4410592 0.4339289 0.4067261 0.4368618 0.4361388 0.4235727
#>  [29] 0.4080665 0.4339289 0.4080665 0.4790668 0.4361388 0.4493858 0.4605267
#>  [36] 0.4998898 0.4470457 0.4567077 0.4601668 0.4185430 0.4529203 0.4158415
#>  [43] 0.4215570 0.4201637 0.4250102 0.4361388 0.4215570 0.4156770 0.4470457
#>  [50] 0.4567077 0.4352432 0.4373813 0.4215772 0.4605267 0.4371012 0.4080665
#>  [57] 0.4397390 0.4740391 0.4998898 0.4304967 0.4215570 0.4359046 0.4250859
#>  [64] 0.4387327 0.4557135 0.4293876 0.4387327 0.4303783 0.4713705 0.4735580
#>  [71] 0.4315458 0.4945374 0.4250926 0.4605267 0.4849514 0.4361388 0.4550573
#>  [78] 0.4207189 0.4529675 0.4304967 0.4605267 0.4456255 0.4700497 0.4758090
#>  [85] 0.4373010 0.4470651 0.4303362 0.4372641 0.4529203 0.4342629 0.4373813
#>  [92] 0.4080665 0.4361388 0.4529203 0.4575762 0.4605267 0.4142942 0.4800401
#>  [99] 0.4207589 0.4308629 0.4185430 0.4219300 0.4361388 0.4215772 0.4232845
#> [106] 0.4515579 0.4758090 0.4156770 0.4371012 0.4219300 0.4174581 0.4156770
#> [113] 0.4493858 0.4215772 0.4628407 0.4529203 0.4628407 0.4493858 0.4082627
#> [120] 0.4567077 0.4813703 0.4067261 0.4158415 0.4758090 0.4215772 0.4315458
#> [127] 0.4550573 0.4168280 0.4567077 0.4250102 0.4156770 0.4363165 0.4372641
#> [134] 0.4250102 0.4339289 0.4308629 0.4303362 0.4067261 0.4606671 0.4372641
#> [141] 0.4262300 0.4605267 0.4185430 0.4601668 0.4609323 0.4232845 0.4217111
#> [148] 0.4387327 0.4372641 0.4098996 0.4213332 0.4584131 0.4217111 0.4567077
#> [155] 0.4512976 0.4734969 0.4142942 0.4529203 0.4235727 0.4266032 0.4228078
#> [162] 0.4569029 0.4185430 0.4219288 0.4606671 0.4082627 0.4232845 0.4308629
#> [169] 0.4151996 0.4207189 0.4080665 0.4614468 0.4174581 0.4082627 0.4219300
#> [176] 0.4363165 0.4550573 0.4377495 0.4550573 0.4219288 0.4611302 0.4813055
#> [183] 0.4335544 0.4395032 0.4377495 0.4168280 0.4614468 0.4228078 0.4142942
#> [190] 0.4585615 0.4142942 0.4319711 0.4377088 0.4510969 0.4270518 0.4737080
#> [197] 0.4293876 0.4373860 0.4232845 0.4232941 0.4610523 0.4185430 0.4286656
#> [204] 0.4361388 0.4425675 0.4158415 0.4518015 0.4373010 0.4342629 0.4225787
#> [211] 0.4231462 0.4266985 0.4611302 0.4067261 0.4614468 0.4609323 0.4250859
#> [218] 0.4493858 0.4361388 0.4207589 0.4396325 0.4098996 0.4285223 0.4235727
#> [225] 0.4250102 0.4372641 0.4470828 0.4231462 0.4475473 0.4364745 0.4174581
#> [232] 0.4168280 0.4325559 0.4339289 0.4321517 0.4234692 0.4283598 0.4207589
#> [239] 0.4304967 0.4321517 0.4584131 0.4368618 0.4339289 0.4228078 0.4363165
#> [246] 0.4067261 0.4219300 0.4207589 0.4185430 0.4482666 0.4270518 0.4304967
#> [253] 0.4605267 0.4387327 0.4359046 0.4584131 0.4151996 0.4319850 0.4606671
#> [260] 0.4361388 0.4607649 0.4286656 0.4773737 0.4098996 0.4377088 0.4550573
#> [267] 0.4604370 0.4225787 0.4142942 0.4458515 0.4361388 0.4168280 0.4325559
#> [274] 0.4207589 0.4428870 0.4715041 0.4738318 0.4319850 0.4369539 0.4758090
#> [281] 0.4219300 0.4575762 0.4361388 0.4605267 0.4602166 0.4221798 0.4151996
#> [288] 0.4158415 0.4225787 0.4321517 0.4377495 0.4235727 0.4368618 0.4168280
#> [295] 0.4285223 0.4207589 0.4185430 0.4352432 0.4401443 0.4814934 0.4207589
#> [302] 0.4368618 0.4605267 0.4731844 0.4321517 0.4595270 0.4507348 0.4232941
#> [309] 0.4528056 0.4531101 0.4614468 0.4168280 0.4628407 0.4313839 0.4207589
#> [316] 0.4303362 0.4201637 0.4215570 0.4158415 0.4201637 0.4215772 0.4228949
#> [323] 0.4082627 0.4082627 0.4142942 0.4142942 0.4067261 0.4080665 0.4205327
#> [330] 0.4215772 0.4082627 0.4142942 0.4080665 0.4080665 0.4232845 0.4215570
#> [337] 0.4067261 0.4142942 0.4215570 0.4264668 0.4082627 0.4228949 0.4174581
#> [344] 0.4156770 0.4080665 0.4156770 0.4232845 0.4082627 0.4174581 0.4142942
#> [351] 0.4215772 0.4080665 0.4232845 0.4080665 0.4156770 0.4067261 0.4207189
#> [358] 0.4067261 0.4232845 0.4262300 0.4363165 0.4361388 0.4266985 0.4485584
#> [365] 0.4648428 0.4584131 0.4228949 0.4067261 0.4067261 0.4213332 0.4758090
#> [372] 0.4609323 0.4228078 0.4225787 0.4219300 0.4142942 0.4321517 0.4493858
#> [379] 0.4082627 0.4215772 0.4080665 0.4628407 0.4387327 0.4602166 0.4363165
#> [386] 0.4131655 0.4080665 0.4372641 0.4300168 0.4080665 0.4304967 0.4339289
#> [393] 0.4493858 0.4456255 0.4339289 0.4207189 0.4362815 0.4373813 0.4080665
#> [400] 0.4158415 0.4321517 0.4550573 0.4606671 0.4151996 0.4735580 0.4232845
#> [407] 0.4156770 0.4067261 0.4475473 0.4280204 0.4377495 0.4415557 0.4361388
#> [414] 0.4232845 0.4303362 0.4228949 0.4431350 0.4372641 0.4387327 0.4232845
#> [421] 0.4363165 0.4602166 0.4354640 0.4609323 0.4321517 0.4493858 0.4213332
#> [428] 0.4715041 0.4266032 0.4304967 0.4250102 0.4283598 0.4567077 0.4300168
#> [435] 0.4293876 0.4371012 0.4207589 0.4363165 0.4151996 0.4518015 0.4207189
#> [442] 0.4363165 0.4300168 0.4285223 0.4151996