Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4485610 0.4287455 0.4684357 0.4460221 0.4424845 0.4287455 0.4657078
#>   [8] 0.4406385 0.4960445 0.4663679 0.4222268 0.4762488 0.4278701 0.4293266
#>  [15] 0.4277554 0.4130456 0.4877140 0.4214835 0.4828923 0.4711471 0.4382680
#>  [22] 0.5101360 0.4446299 0.4397002 0.4114030 0.4427396 0.4428111 0.4293266
#>  [29] 0.4130147 0.4397002 0.4130147 0.4856605 0.4428111 0.4578314 0.4699051
#>  [36] 0.5124726 0.4542332 0.4657078 0.4691771 0.4229120 0.4611877 0.4222268
#>  [43] 0.4296168 0.4278701 0.4306735 0.4428111 0.4296168 0.4222345 0.4542332
#>  [50] 0.4657078 0.4422607 0.4417374 0.4274268 0.4699051 0.4426350 0.4130147
#>  [57] 0.4429977 0.4804690 0.5124726 0.4361763 0.4296168 0.4431992 0.4265237
#>  [64] 0.4460221 0.4622162 0.4353356 0.4460221 0.4355725 0.4784461 0.4839142
#>  [71] 0.4390326 0.5058447 0.4306279 0.4699051 0.4987830 0.4428111 0.4636047
#>  [78] 0.4265445 0.4580172 0.4361763 0.4699051 0.4541231 0.4753350 0.4877140
#>  [85] 0.4459705 0.4503169 0.4375994 0.4443909 0.4611877 0.4422745 0.4417374
#>  [92] 0.4130147 0.4428111 0.4611877 0.4640960 0.4699051 0.4205648 0.4881273
#>  [99] 0.4259289 0.4358978 0.4229120 0.4276204 0.4428111 0.4274268 0.4291316
#> [106] 0.4584412 0.4877140 0.4222345 0.4426350 0.4276204 0.4234663 0.4222345
#> [113] 0.4578314 0.4274268 0.4725120 0.4611877 0.4725120 0.4578314 0.4130456
#> [120] 0.4657078 0.4942183 0.4114030 0.4222268 0.4877140 0.4274268 0.4390326
#> [127] 0.4636047 0.4214835 0.4657078 0.4306735 0.4222345 0.4418481 0.4443909
#> [134] 0.4306735 0.4397002 0.4358978 0.4375994 0.4114030 0.4686518 0.4443909
#> [141] 0.4322443 0.4699051 0.4229120 0.4691771 0.4673368 0.4291316 0.4294644
#> [148] 0.4460221 0.4443909 0.4143452 0.4255348 0.4669489 0.4294644 0.4657078
#> [155] 0.4579548 0.4851573 0.4205648 0.4611877 0.4293266 0.4335410 0.4287455
#> [162] 0.4621262 0.4229120 0.4264133 0.4686518 0.4130456 0.4291316 0.4358978
#> [169] 0.4195037 0.4265445 0.4130147 0.4711471 0.4234663 0.4130456 0.4276204
#> [176] 0.4418481 0.4636047 0.4410434 0.4636047 0.4264133 0.4663679 0.4953833
#> [183] 0.4371124 0.4441959 0.4410434 0.4214835 0.4711471 0.4287455 0.4205648
#> [190] 0.4648526 0.4205648 0.4390667 0.4399575 0.4575692 0.4284182 0.4864715
#> [197] 0.4353356 0.4439559 0.4291316 0.4306387 0.4651466 0.4229120 0.4356973
#> [204] 0.4428111 0.4463965 0.4222268 0.4587177 0.4459705 0.4422745 0.4255303
#> [211] 0.4239952 0.4315585 0.4663679 0.4114030 0.4711471 0.4673368 0.4265237
#> [218] 0.4578314 0.4428111 0.4259289 0.4425074 0.4143452 0.4353624 0.4293266
#> [225] 0.4306735 0.4443909 0.4516770 0.4239952 0.4561378 0.4419230 0.4234663
#> [232] 0.4214835 0.4340973 0.4397002 0.4382680 0.4286490 0.4338627 0.4259289
#> [239] 0.4361763 0.4382680 0.4669489 0.4427396 0.4397002 0.4287455 0.4418481
#> [246] 0.4114030 0.4276204 0.4259289 0.4229120 0.4561289 0.4284182 0.4361763
#> [253] 0.4699051 0.4460221 0.4431992 0.4669489 0.4195037 0.4370159 0.4686518
#> [260] 0.4428111 0.4684357 0.4356973 0.4784972 0.4143452 0.4399575 0.4636047
#> [267] 0.4660519 0.4255303 0.4205648 0.4475406 0.4428111 0.4214835 0.4340973
#> [274] 0.4259289 0.4474121 0.4769820 0.4804439 0.4370159 0.4406385 0.4877140
#> [281] 0.4276204 0.4640960 0.4428111 0.4699051 0.4679331 0.4245118 0.4195037
#> [288] 0.4222268 0.4255303 0.4382680 0.4410434 0.4293266 0.4427396 0.4214835
#> [295] 0.4353624 0.4259289 0.4229120 0.4422607 0.4467465 0.4940254 0.4259289
#> [302] 0.4427396 0.4699051 0.4832871 0.4382680 0.4603210 0.4548648 0.4306387
#> [309] 0.4595295 0.4562790 0.4711471 0.4214835 0.4725120 0.4354235 0.4259289
#> [316] 0.4375994 0.4278701 0.4296168 0.4222268 0.4278701 0.4274268 0.4289576
#> [323] 0.4130456 0.4130456 0.4205648 0.4205648 0.4114030 0.4130147 0.4260278
#> [330] 0.4274268 0.4130456 0.4205648 0.4130147 0.4130147 0.4291316 0.4296168
#> [337] 0.4114030 0.4205648 0.4296168 0.4321015 0.4130456 0.4289576 0.4234663
#> [344] 0.4222345 0.4130147 0.4222345 0.4291316 0.4130456 0.4234663 0.4205648
#> [351] 0.4274268 0.4130147 0.4291316 0.4130147 0.4222345 0.4114030 0.4265445
#> [358] 0.4114030 0.4291316 0.4322443 0.4418481 0.4428111 0.4315585 0.4520305
#> [365] 0.4695329 0.4669489 0.4289576 0.4114030 0.4114030 0.4255348 0.4877140
#> [372] 0.4673368 0.4287455 0.4255303 0.4276204 0.4205648 0.4382680 0.4578314
#> [379] 0.4130456 0.4274268 0.4130147 0.4725120 0.4460221 0.4679331 0.4418481
#> [386] 0.4174134 0.4130147 0.4443909 0.4362458 0.4130147 0.4361763 0.4397002
#> [393] 0.4578314 0.4541231 0.4397002 0.4265445 0.4432716 0.4417374 0.4130147
#> [400] 0.4222268 0.4382680 0.4636047 0.4686518 0.4195037 0.4839142 0.4291316
#> [407] 0.4222345 0.4114030 0.4561378 0.4332454 0.4410434 0.4444258 0.4428111
#> [414] 0.4291316 0.4375994 0.4289576 0.4479225 0.4443909 0.4460221 0.4291316
#> [421] 0.4418481 0.4679331 0.4441300 0.4673368 0.4382680 0.4578314 0.4255348
#> [428] 0.4769820 0.4335410 0.4361763 0.4306735 0.4338627 0.4657078 0.4362458
#> [435] 0.4353356 0.4426350 0.4259289 0.4418481 0.4195037 0.4587177 0.4265445
#> [442] 0.4418481 0.4362458 0.4353624 0.4195037