Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4519540 0.4340426 0.4648118 0.4468874 0.4461573 0.4340426 0.4617979
#>   [8] 0.4444103 0.4855697 0.4652994 0.4291721 0.4703495 0.4333054 0.4341722
#>  [15] 0.4352183 0.4218786 0.4787722 0.4275632 0.4782423 0.4661796 0.4402819
#>  [22] 0.4978487 0.4484971 0.4415217 0.4209371 0.4444282 0.4439738 0.4341722
#>  [29] 0.4221550 0.4415217 0.4221550 0.4857269 0.4439738 0.4565566 0.4651960
#>  [36] 0.5000168 0.4570567 0.4617979 0.4674481 0.4287116 0.4591559 0.4291721
#>  [43] 0.4345848 0.4333054 0.4354049 0.4439738 0.4345848 0.4294335 0.4570567
#>  [50] 0.4617979 0.4442565 0.4446935 0.4332137 0.4651960 0.4445244 0.4221550
#>  [57] 0.4474559 0.4808151 0.5000168 0.4393878 0.4345848 0.4449201 0.4346245
#>  [64] 0.4468874 0.4643170 0.4389208 0.4468874 0.4404869 0.4789432 0.4760020
#>  [71] 0.4416530 0.4949888 0.4386081 0.4651960 0.4869659 0.4439738 0.4608567
#>  [78] 0.4324185 0.4586852 0.4393878 0.4651960 0.4541544 0.4726580 0.4787722
#>  [85] 0.4467082 0.4529736 0.4400365 0.4450665 0.4591559 0.4436233 0.4446935
#>  [92] 0.4221550 0.4439738 0.4591559 0.4631951 0.4651960 0.4281768 0.4823447
#>  [99] 0.4311527 0.4401006 0.4287116 0.4322762 0.4439738 0.4332137 0.4342494
#> [106] 0.4588844 0.4787722 0.4294335 0.4445244 0.4322762 0.4301129 0.4294335
#> [113] 0.4565566 0.4332137 0.4676637 0.4591559 0.4676637 0.4565566 0.4218786
#> [120] 0.4617979 0.4831960 0.4209371 0.4291721 0.4787722 0.4332137 0.4416530
#> [127] 0.4608567 0.4275632 0.4617979 0.4354049 0.4294335 0.4438135 0.4450665
#> [134] 0.4354049 0.4415217 0.4401006 0.4400365 0.4209371 0.4648942 0.4450665
#> [141] 0.4360679 0.4651960 0.4287116 0.4674481 0.4649144 0.4342494 0.4342868
#> [148] 0.4468874 0.4450665 0.4229017 0.4313693 0.4628342 0.4342868 0.4617979
#> [155] 0.4586987 0.4763838 0.4281768 0.4591559 0.4341722 0.4374716 0.4340426
#> [162] 0.4651345 0.4287116 0.4319634 0.4648942 0.4218786 0.4342494 0.4401006
#> [169] 0.4267094 0.4324185 0.4221550 0.4661796 0.4301129 0.4218786 0.4322762
#> [176] 0.4438135 0.4608567 0.4452266 0.4608567 0.4319634 0.4652994 0.4835900
#> [183] 0.4424883 0.4480544 0.4452266 0.4275632 0.4661796 0.4340426 0.4281768
#> [190] 0.4643623 0.4281768 0.4411434 0.4450671 0.4603791 0.4362066 0.4774025
#> [197] 0.4389208 0.4472250 0.4342494 0.4351573 0.4649399 0.4287116 0.4391351
#> [204] 0.4439738 0.4498672 0.4291721 0.4590363 0.4467082 0.4436233 0.4323546
#> [211] 0.4328484 0.4378543 0.4652994 0.4209371 0.4661796 0.4649144 0.4346245
#> [218] 0.4565566 0.4439738 0.4311527 0.4468474 0.4229017 0.4385270 0.4341722
#> [225] 0.4354049 0.4450665 0.4541358 0.4328484 0.4552604 0.4452711 0.4301129
#> [232] 0.4275632 0.4406861 0.4415217 0.4402819 0.4356114 0.4378160 0.4311527
#> [239] 0.4393878 0.4402819 0.4628342 0.4444282 0.4415217 0.4340426 0.4438135
#> [246] 0.4209371 0.4322762 0.4311527 0.4287116 0.4580746 0.4362066 0.4393878
#> [253] 0.4651960 0.4468874 0.4449201 0.4628342 0.4267094 0.4416251 0.4648942
#> [260] 0.4439738 0.4648118 0.4391351 0.4812792 0.4229017 0.4450671 0.4608567
#> [267] 0.4645675 0.4323546 0.4281768 0.4528584 0.4439738 0.4275632 0.4406861
#> [274] 0.4311527 0.4508311 0.4738392 0.4759890 0.4416251 0.4444103 0.4787722
#> [281] 0.4322762 0.4631951 0.4439738 0.4651960 0.4644311 0.4320728 0.4267094
#> [288] 0.4291721 0.4323546 0.4402819 0.4452266 0.4341722 0.4444282 0.4275632
#> [295] 0.4385270 0.4311527 0.4287116 0.4442565 0.4498050 0.4830887 0.4311527
#> [302] 0.4444282 0.4651960 0.4756039 0.4402819 0.4629433 0.4597174 0.4351573
#> [309] 0.4601873 0.4608976 0.4661796 0.4275632 0.4676637 0.4406893 0.4311527
#> [316] 0.4400365 0.4333054 0.4345848 0.4291721 0.4333054 0.4332137 0.4343569
#> [323] 0.4218786 0.4218786 0.4281768 0.4281768 0.4209371 0.4221550 0.4338661
#> [330] 0.4332137 0.4218786 0.4281768 0.4221550 0.4221550 0.4342494 0.4345848
#> [337] 0.4209371 0.4281768 0.4345848 0.4394519 0.4218786 0.4343569 0.4301129
#> [344] 0.4294335 0.4221550 0.4294335 0.4342494 0.4218786 0.4301129 0.4281768
#> [351] 0.4332137 0.4221550 0.4342494 0.4221550 0.4294335 0.4209371 0.4324185
#> [358] 0.4209371 0.4342494 0.4360679 0.4438135 0.4439738 0.4378543 0.4542816
#> [365] 0.4681381 0.4628342 0.4343569 0.4209371 0.4209371 0.4313693 0.4787722
#> [372] 0.4649144 0.4340426 0.4323546 0.4322762 0.4281768 0.4402819 0.4565566
#> [379] 0.4218786 0.4332137 0.4221550 0.4676637 0.4468874 0.4644311 0.4438135
#> [386] 0.4251508 0.4221550 0.4450665 0.4395319 0.4221550 0.4393878 0.4415217
#> [393] 0.4565566 0.4541544 0.4415217 0.4324185 0.4450369 0.4446935 0.4221550
#> [400] 0.4291721 0.4402819 0.4608567 0.4648942 0.4267094 0.4760020 0.4342494
#> [407] 0.4294335 0.4209371 0.4552604 0.4403107 0.4452266 0.4484246 0.4439738
#> [414] 0.4342494 0.4400365 0.4343569 0.4510737 0.4450665 0.4468874 0.4342494
#> [421] 0.4438135 0.4644311 0.4446903 0.4649144 0.4402819 0.4565566 0.4313693
#> [428] 0.4738392 0.4374716 0.4393878 0.4354049 0.4378160 0.4617979 0.4395319
#> [435] 0.4389208 0.4445244 0.4311527 0.4438135 0.4267094 0.4590363 0.4324185
#> [442] 0.4438135 0.4395319 0.4385270 0.4267094