Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4375855 0.4149828 0.4582973 0.4331089 0.4312241 0.4149828 0.4535006
#>   [8] 0.4297621 0.4828424 0.4582894 0.4080682 0.4634624 0.4126473 0.4160155
#>  [15] 0.4135507 0.3984058 0.4765275 0.4071387 0.4744306 0.4598024 0.4242190
#>  [22] 0.5013981 0.4337585 0.4260421 0.3966234 0.4304647 0.4293553 0.4160155
#>  [29] 0.3977671 0.4260421 0.3977671 0.4808234 0.4293553 0.4455225 0.4584250
#>  [36] 0.5042815 0.4452356 0.4535006 0.4579923 0.4089816 0.4488491 0.4080682
#>  [43] 0.4137877 0.4126473 0.4166515 0.4293553 0.4137877 0.4074083 0.4452356
#>  [50] 0.4535006 0.4295371 0.4304537 0.4130834 0.4584250 0.4302666 0.3977671
#>  [57] 0.4321498 0.4749686 0.5042815 0.4222425 0.4137877 0.4304783 0.4154798
#>  [64] 0.4331089 0.4549685 0.4230487 0.4331089 0.4228652 0.4719856 0.4734126
#>  [71] 0.4262663 0.4971843 0.4191476 0.4584250 0.4873706 0.4293553 0.4514822
#>  [78] 0.4124437 0.4472241 0.4222425 0.4584250 0.4416233 0.4666109 0.4765275
#>  [85] 0.4326325 0.4391988 0.4232109 0.4308710 0.4488491 0.4284085 0.4304537
#>  [92] 0.3977671 0.4293553 0.4488491 0.4536236 0.4584250 0.4062730 0.4804377
#>  [99] 0.4121659 0.4243269 0.4089816 0.4137247 0.4293553 0.4130834 0.4150245
#> [106] 0.4479511 0.4765275 0.4074083 0.4302666 0.4137247 0.4094618 0.4074083
#> [113] 0.4455225 0.4130834 0.4618274 0.4488491 0.4618274 0.4455225 0.3984058
#> [120] 0.4535006 0.4834261 0.3966234 0.4080682 0.4765275 0.4130834 0.4262663
#> [127] 0.4514822 0.4071387 0.4535006 0.4166515 0.4074083 0.4296896 0.4308710
#> [134] 0.4166515 0.4260421 0.4243269 0.4232109 0.3966234 0.4587858 0.4308710
#> [141] 0.4188794 0.4584250 0.4089816 0.4579923 0.4583958 0.4150245 0.4143543
#> [148] 0.4331089 0.4308710 0.3998870 0.4127514 0.4552304 0.4143543 0.4535006
#> [155] 0.4478559 0.4731809 0.4062730 0.4488491 0.4160155 0.4185742 0.4149828
#> [162] 0.4556437 0.4089816 0.4135478 0.4587858 0.3984058 0.4150245 0.4243269
#> [169] 0.4051822 0.4124437 0.3977671 0.4598024 0.4094618 0.3984058 0.4137247
#> [176] 0.4296896 0.4514822 0.4304701 0.4514822 0.4135478 0.4582894 0.4831106
#> [183] 0.4262971 0.4329704 0.4304701 0.4071387 0.4598024 0.4149828 0.4062730
#> [190] 0.4544832 0.4062730 0.4250583 0.4297115 0.4494332 0.4174026 0.4734226
#> [197] 0.4230487 0.4322095 0.4150245 0.4156621 0.4574157 0.4089816 0.4211319
#> [204] 0.4293553 0.4362253 0.4080682 0.4482241 0.4326325 0.4284085 0.4137258
#> [211] 0.4133082 0.4182186 0.4582894 0.3966234 0.4598024 0.4583958 0.4154798
#> [218] 0.4455225 0.4293553 0.4121659 0.4317879 0.3998870 0.4221556 0.4160155
#> [225] 0.4166515 0.4308710 0.4403364 0.4133082 0.4436942 0.4304353 0.4094618
#> [232] 0.4071387 0.4228286 0.4260421 0.4242190 0.4153442 0.4197745 0.4121659
#> [239] 0.4222425 0.4242190 0.4552304 0.4304647 0.4260421 0.4149828 0.4296896
#> [246] 0.3966234 0.4137247 0.4121659 0.4089816 0.4467097 0.4174026 0.4222425
#> [253] 0.4584250 0.4331089 0.4304783 0.4552304 0.4051822 0.4247247 0.4587858
#> [260] 0.4293553 0.4582973 0.4211319 0.4778102 0.3998870 0.4297115 0.4514822
#> [267] 0.4576714 0.4137258 0.4062730 0.4396629 0.4293553 0.4071387 0.4228286
#> [274] 0.4121659 0.4368357 0.4689729 0.4719982 0.4247247 0.4297621 0.4765275
#> [281] 0.4137247 0.4536236 0.4293553 0.4584250 0.4580573 0.4131295 0.4051822
#> [288] 0.4080682 0.4137258 0.4242190 0.4304701 0.4160155 0.4304647 0.4071387
#> [295] 0.4221556 0.4121659 0.4089816 0.4295371 0.4362208 0.4828342 0.4121659
#> [302] 0.4304647 0.4584250 0.4727562 0.4242190 0.4548207 0.4480590 0.4156621
#> [309] 0.4491539 0.4492273 0.4598024 0.4071387 0.4618274 0.4245780 0.4121659
#> [316] 0.4232109 0.4126473 0.4137877 0.4080682 0.4126473 0.4130834 0.4140415
#> [323] 0.3984058 0.3984058 0.4062730 0.4062730 0.3966234 0.3977671 0.4123847
#> [330] 0.4130834 0.3984058 0.4062730 0.3977671 0.3977671 0.4150245 0.4137877
#> [337] 0.3966234 0.4062730 0.4137877 0.4207941 0.3984058 0.4140415 0.4094618
#> [344] 0.4074083 0.3977671 0.4074083 0.4150245 0.3984058 0.4094618 0.4062730
#> [351] 0.4130834 0.3977671 0.4150245 0.3977671 0.4074083 0.3966234 0.4124437
#> [358] 0.3966234 0.4150245 0.4188794 0.4296896 0.4293553 0.4182186 0.4416645
#> [365] 0.4612382 0.4552304 0.4140415 0.3966234 0.3966234 0.4127514 0.4765275
#> [372] 0.4583958 0.4149828 0.4137258 0.4137247 0.4062730 0.4242190 0.4455225
#> [379] 0.3984058 0.4130834 0.3977671 0.4618274 0.4331089 0.4580573 0.4296896
#> [386] 0.4028345 0.3977671 0.4308710 0.4239446 0.3977671 0.4222425 0.4260421
#> [393] 0.4455225 0.4416233 0.4260421 0.4124437 0.4304806 0.4304537 0.3977671
#> [400] 0.4080682 0.4242190 0.4514822 0.4587858 0.4051822 0.4734126 0.4150245
#> [407] 0.4074083 0.3966234 0.4436942 0.4220830 0.4304701 0.4336531 0.4293553
#> [414] 0.4150245 0.4232109 0.4140415 0.4369692 0.4308710 0.4331089 0.4150245
#> [421] 0.4296896 0.4580573 0.4300603 0.4583958 0.4242190 0.4455225 0.4127514
#> [428] 0.4689729 0.4185742 0.4222425 0.4166515 0.4197745 0.4535006 0.4239446
#> [435] 0.4230487 0.4302666 0.4121659 0.4296896 0.4051822 0.4482241 0.4124437
#> [442] 0.4296896 0.4239446 0.4221556 0.4051822
# }