Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4430033 0.4201991 0.4546018 0.4353382 0.4354888 0.4201991 0.4485395
#>   [8] 0.4357527 0.4715580 0.4556421 0.4145322 0.4564952 0.4190120 0.4228895
#>  [15] 0.4210583 0.4093757 0.4662630 0.4163254 0.4685971 0.4534757 0.4288975
#>  [22] 0.4883417 0.4378243 0.4302719 0.4083427 0.4347080 0.4326206 0.4228895
#>  [29] 0.4091377 0.4302719 0.4091377 0.4741037 0.4326206 0.4423350 0.4522118
#>  [36] 0.4900951 0.4437650 0.4485395 0.4565286 0.4177243 0.4452071 0.4145322
#>  [43] 0.4198536 0.4190120 0.4227932 0.4326206 0.4198536 0.4143040 0.4437650
#>  [50] 0.4485395 0.4326205 0.4357731 0.4206661 0.4522118 0.4353539 0.4091377
#>  [57] 0.4365287 0.4683855 0.4900951 0.4275371 0.4198536 0.4332687 0.4261034
#>  [64] 0.4353382 0.4515356 0.4271131 0.4353382 0.4280045 0.4676164 0.4654843
#>  [71] 0.4281908 0.4859663 0.4252635 0.4522118 0.4759766 0.4326206 0.4472146
#>  [78] 0.4181705 0.4498280 0.4275371 0.4522118 0.4399872 0.4631331 0.4662630
#>  [85] 0.4335845 0.4442244 0.4269333 0.4339780 0.4452071 0.4306937 0.4357731
#>  [92] 0.4091377 0.4326206 0.4452071 0.4547368 0.4522118 0.4134895 0.4723952
#>  [99] 0.4200310 0.4291066 0.4177243 0.4213996 0.4326206 0.4206661 0.4217510
#> [106] 0.4485247 0.4662630 0.4143040 0.4353539 0.4213996 0.4154770 0.4143040
#> [113] 0.4423350 0.4206661 0.4548210 0.4452071 0.4548210 0.4423350 0.4093757
#> [120] 0.4485395 0.4727708 0.4083427 0.4145322 0.4662630 0.4206661 0.4281908
#> [127] 0.4472146 0.4163254 0.4485395 0.4227932 0.4143040 0.4340666 0.4339780
#> [134] 0.4227932 0.4302719 0.4291066 0.4269333 0.4083427 0.4540858 0.4339780
#> [141] 0.4229214 0.4522118 0.4177243 0.4565286 0.4549153 0.4217510 0.4199621
#> [148] 0.4353382 0.4339780 0.4103548 0.4217036 0.4498781 0.4199621 0.4485395
#> [155] 0.4475599 0.4636728 0.4134895 0.4452071 0.4228895 0.4235647 0.4201991
#> [162] 0.4529232 0.4177243 0.4223463 0.4540858 0.4093757 0.4217510 0.4291066
#> [169] 0.4150069 0.4181705 0.4091377 0.4534757 0.4154770 0.4093757 0.4213996
#> [176] 0.4340666 0.4472146 0.4365662 0.4472146 0.4223463 0.4556421 0.4708392
#> [183] 0.4318388 0.4373645 0.4365662 0.4163254 0.4534757 0.4201991 0.4134895
#> [190] 0.4558432 0.4134895 0.4282653 0.4367698 0.4466802 0.4278590 0.4636944
#> [197] 0.4271131 0.4346012 0.4217510 0.4208922 0.4559053 0.4177243 0.4256140
#> [204] 0.4326206 0.4395771 0.4145322 0.4486618 0.4335845 0.4306937 0.4234407
#> [211] 0.4245983 0.4245955 0.4556421 0.4083427 0.4534757 0.4549153 0.4261034
#> [218] 0.4423350 0.4326206 0.4200310 0.4382510 0.4103548 0.4252913 0.4228895
#> [225] 0.4227932 0.4339780 0.4433336 0.4245983 0.4411978 0.4346302 0.4154770
#> [232] 0.4163254 0.4320893 0.4302719 0.4288975 0.4219489 0.4255578 0.4200310
#> [239] 0.4275371 0.4288975 0.4498781 0.4347080 0.4302719 0.4201991 0.4340666
#> [246] 0.4083427 0.4213996 0.4200310 0.4177243 0.4450414 0.4278590 0.4275371
#> [253] 0.4522118 0.4353382 0.4332687 0.4498781 0.4150069 0.4293102 0.4540858
#> [260] 0.4326206 0.4546018 0.4256140 0.4755314 0.4103548 0.4367698 0.4472146
#> [267] 0.4549102 0.4234407 0.4134895 0.4457129 0.4326206 0.4163254 0.4320893
#> [274] 0.4200310 0.4417818 0.4647340 0.4671437 0.4293102 0.4357527 0.4662630
#> [281] 0.4213996 0.4547368 0.4326206 0.4522118 0.4534687 0.4234516 0.4150069
#> [288] 0.4145322 0.4234407 0.4288975 0.4365662 0.4228895 0.4347080 0.4163254
#> [295] 0.4252913 0.4200310 0.4177243 0.4326205 0.4379692 0.4733056 0.4200310
#> [302] 0.4347080 0.4522118 0.4648843 0.4288975 0.4579121 0.4484780 0.4208922
#> [309] 0.4497944 0.4510557 0.4534757 0.4163254 0.4548210 0.4300427 0.4200310
#> [316] 0.4269333 0.4190120 0.4198536 0.4145322 0.4190120 0.4206661 0.4213515
#> [323] 0.4093757 0.4093757 0.4134895 0.4134895 0.4083427 0.4091377 0.4201967
#> [330] 0.4206661 0.4093757 0.4134895 0.4091377 0.4091377 0.4217510 0.4198536
#> [337] 0.4083427 0.4134895 0.4198536 0.4260879 0.4093757 0.4213515 0.4154770
#> [344] 0.4143040 0.4091377 0.4143040 0.4217510 0.4093757 0.4154770 0.4134895
#> [351] 0.4206661 0.4091377 0.4217510 0.4091377 0.4143040 0.4083427 0.4181705
#> [358] 0.4083427 0.4217510 0.4229214 0.4340666 0.4326206 0.4245955 0.4459322
#> [365] 0.4589654 0.4498781 0.4213515 0.4083427 0.4083427 0.4217036 0.4662630
#> [372] 0.4549153 0.4201991 0.4234407 0.4213996 0.4134895 0.4288975 0.4423350
#> [379] 0.4093757 0.4206661 0.4091377 0.4548210 0.4353382 0.4534687 0.4340666
#> [386] 0.4130266 0.4091377 0.4339780 0.4277850 0.4091377 0.4275371 0.4302719
#> [393] 0.4423350 0.4399872 0.4302719 0.4181705 0.4340681 0.4357731 0.4091377
#> [400] 0.4145322 0.4288975 0.4472146 0.4540858 0.4150069 0.4654843 0.4217510
#> [407] 0.4143040 0.4083427 0.4411978 0.4269350 0.4365662 0.4399957 0.4326206
#> [414] 0.4217510 0.4269333 0.4213515 0.4427137 0.4339780 0.4353382 0.4217510
#> [421] 0.4340666 0.4534687 0.4320144 0.4549153 0.4288975 0.4423350 0.4217036
#> [428] 0.4647340 0.4235647 0.4275371 0.4227932 0.4255578 0.4485395 0.4277850
#> [435] 0.4271131 0.4353539 0.4200310 0.4340666 0.4150069 0.4486618 0.4181705
#> [442] 0.4340666 0.4277850 0.4252913 0.4150069