Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4408988 0.4170366 0.4579335 0.4338686 0.4324653 0.4170366 0.4535466
#>   [8] 0.4321443 0.4798552 0.4588935 0.4099476 0.4636780 0.4147998 0.4174044
#>  [15] 0.4166241 0.4010555 0.4744167 0.4097071 0.4746670 0.4589949 0.4262370
#>  [22] 0.4989381 0.4368260 0.4276550 0.3999606 0.4314598 0.4304375 0.4174044
#>  [29] 0.4012900 0.4276550 0.4012900 0.4798083 0.4304375 0.4465467 0.4575742
#>  [36] 0.5011270 0.4445465 0.4535466 0.4583723 0.4111070 0.4500097 0.4099476
#>  [43] 0.4162335 0.4147998 0.4193820 0.4304375 0.4162335 0.4102088 0.4445465
#>  [50] 0.4535466 0.4296120 0.4327245 0.4161405 0.4575742 0.4317186 0.4012900
#>  [57] 0.4346394 0.4741668 0.5011270 0.4246962 0.4162335 0.4303651 0.4187762
#>  [64] 0.4338686 0.4536235 0.4236642 0.4338686 0.4243184 0.4719000 0.4716924
#>  [71] 0.4263977 0.4959307 0.4213609 0.4575742 0.4830321 0.4304375 0.4520156
#>  [78] 0.4150545 0.4490727 0.4246962 0.4575742 0.4429515 0.4688017 0.4744167
#>  [85] 0.4322402 0.4430599 0.4241500 0.4321034 0.4500097 0.4282676 0.4327245
#>  [92] 0.4012900 0.4304375 0.4500097 0.4551096 0.4575742 0.4088037 0.4795814
#>  [99] 0.4138039 0.4259279 0.4111070 0.4154475 0.4304375 0.4161405 0.4172576
#> [106] 0.4488399 0.4744167 0.4102088 0.4317186 0.4154475 0.4117631 0.4102088
#> [113] 0.4465467 0.4161405 0.4605901 0.4500097 0.4605901 0.4465467 0.4010555
#> [120] 0.4535466 0.4803065 0.3999606 0.4099476 0.4744167 0.4161405 0.4263977
#> [127] 0.4520156 0.4097071 0.4535466 0.4193820 0.4102088 0.4307613 0.4321034
#> [134] 0.4193820 0.4276550 0.4259279 0.4241500 0.3999606 0.4578324 0.4321034
#> [141] 0.4198771 0.4575742 0.4111070 0.4583723 0.4588715 0.4172576 0.4157835
#> [148] 0.4338686 0.4321034 0.4030148 0.4145488 0.4548578 0.4157835 0.4535466
#> [155] 0.4481588 0.4713802 0.4088037 0.4500097 0.4174044 0.4207339 0.4170366
#> [162] 0.4550926 0.4111070 0.4152780 0.4578324 0.4010555 0.4172576 0.4259279
#> [169] 0.4082109 0.4150545 0.4012900 0.4589949 0.4117631 0.4010555 0.4154475
#> [176] 0.4307613 0.4520156 0.4328343 0.4520156 0.4152780 0.4588935 0.4798096
#> [183] 0.4283142 0.4345443 0.4328343 0.4097071 0.4589949 0.4170366 0.4088037
#> [190] 0.4563055 0.4088037 0.4254064 0.4326610 0.4484507 0.4206490 0.4720578
#> [197] 0.4236642 0.4323318 0.4172576 0.4174829 0.4586293 0.4111070 0.4227025
#> [204] 0.4304375 0.4388835 0.4099476 0.4491248 0.4322402 0.4282676 0.4166617
#> [211] 0.4167772 0.4208987 0.4588935 0.3999606 0.4589949 0.4588715 0.4187762
#> [218] 0.4465467 0.4304375 0.4138039 0.4346638 0.4030148 0.4226621 0.4174044
#> [225] 0.4193820 0.4321034 0.4430022 0.4167772 0.4442760 0.4315828 0.4117631
#> [232] 0.4097071 0.4271924 0.4276550 0.4262370 0.4176719 0.4227634 0.4138039
#> [239] 0.4246962 0.4262370 0.4548578 0.4314598 0.4276550 0.4170366 0.4307613
#> [246] 0.3999606 0.4154475 0.4138039 0.4111070 0.4463132 0.4206490 0.4246962
#> [253] 0.4575742 0.4338686 0.4303651 0.4548578 0.4082109 0.4255444 0.4578324
#> [260] 0.4304375 0.4579335 0.4227025 0.4764069 0.4030148 0.4326610 0.4520156
#> [267] 0.4582949 0.4166617 0.4088037 0.4415938 0.4304375 0.4097071 0.4271924
#> [274] 0.4138039 0.4385362 0.4706816 0.4724046 0.4255444 0.4321443 0.4744167
#> [281] 0.4154475 0.4551096 0.4304375 0.4575742 0.4572984 0.4161052 0.4082109
#> [288] 0.4099476 0.4166617 0.4262370 0.4328343 0.4174044 0.4314598 0.4097071
#> [295] 0.4226621 0.4138039 0.4111070 0.4296120 0.4363209 0.4803229 0.4138039
#> [302] 0.4314598 0.4575742 0.4712012 0.4262370 0.4554656 0.4486274 0.4174829
#> [309] 0.4504011 0.4512909 0.4589949 0.4097071 0.4605901 0.4263281 0.4138039
#> [316] 0.4241500 0.4147998 0.4162335 0.4099476 0.4147998 0.4161405 0.4173664
#> [323] 0.4010555 0.4010555 0.4088037 0.4088037 0.3999606 0.4012900 0.4150982
#> [330] 0.4161405 0.4010555 0.4088037 0.4012900 0.4012900 0.4172576 0.4162335
#> [337] 0.3999606 0.4088037 0.4162335 0.4222424 0.4010555 0.4173664 0.4117631
#> [344] 0.4102088 0.4012900 0.4102088 0.4172576 0.4010555 0.4117631 0.4088037
#> [351] 0.4161405 0.4012900 0.4172576 0.4012900 0.4102088 0.3999606 0.4150545
#> [358] 0.3999606 0.4172576 0.4198771 0.4307613 0.4304375 0.4208987 0.4450192
#> [365] 0.4622885 0.4548578 0.4173664 0.3999606 0.3999606 0.4145488 0.4744167
#> [372] 0.4588715 0.4170366 0.4166617 0.4154475 0.4088037 0.4262370 0.4465467
#> [379] 0.4010555 0.4161405 0.4012900 0.4605901 0.4338686 0.4572984 0.4307613
#> [386] 0.4062756 0.4012900 0.4321034 0.4244064 0.4012900 0.4246962 0.4276550
#> [393] 0.4465467 0.4429515 0.4276550 0.4150545 0.4307436 0.4327245 0.4012900
#> [400] 0.4099476 0.4262370 0.4520156 0.4578324 0.4082109 0.4716924 0.4172576
#> [407] 0.4102088 0.3999606 0.4442760 0.4239113 0.4328343 0.4365004 0.4304375
#> [414] 0.4172576 0.4241500 0.4173664 0.4392738 0.4321034 0.4338686 0.4172576
#> [421] 0.4307613 0.4572984 0.4300412 0.4588715 0.4262370 0.4465467 0.4145488
#> [428] 0.4706816 0.4207339 0.4246962 0.4193820 0.4227634 0.4535466 0.4244064
#> [435] 0.4236642 0.4317186 0.4138039 0.4307613 0.4082109 0.4491248 0.4150545
#> [442] 0.4307613 0.4244064 0.4226621 0.4082109