Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
plotBart:::propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4442050 0.4217379 0.4621295 0.4397252 0.4368219 0.4217379 0.4587111
#>   [8] 0.4349013 0.4891639 0.4618652 0.4157524 0.4704034 0.4209158 0.4224480
#>  [15] 0.4212944 0.4062517 0.4816789 0.4141876 0.4790964 0.4649521 0.4314892
#>  [22] 0.5056294 0.4404706 0.4333050 0.4047713 0.4359092 0.4363298 0.4224480
#>  [29] 0.4060208 0.4333050 0.4060208 0.4845752 0.4363298 0.4511944 0.4635348
#>  [36] 0.5077828 0.4502719 0.4587111 0.4630375 0.4159352 0.4552538 0.4157524
#>  [43] 0.4222715 0.4209158 0.4242620 0.4363298 0.4222715 0.4155394 0.4502719
#>  [50] 0.4587111 0.4350235 0.4355806 0.4216600 0.4635348 0.4357103 0.4060208
#>  [57] 0.4386323 0.4796432 0.5077828 0.4294843 0.4222715 0.4359772 0.4213793
#>  [64] 0.4397252 0.4593098 0.4285069 0.4397252 0.4289106 0.4759749 0.4777809
#>  [71] 0.4324049 0.5006796 0.4262471 0.4635348 0.4920541 0.4363298 0.4567567
#>  [78] 0.4203349 0.4537534 0.4294843 0.4635348 0.4481775 0.4716618 0.4816789
#>  [85] 0.4388243 0.4459941 0.4301605 0.4379200 0.4552538 0.4350004 0.4355806
#>  [92] 0.4060208 0.4363298 0.4552538 0.4578834 0.4635348 0.4141965 0.4834591
#>  [99] 0.4188853 0.4295286 0.4159352 0.4204575 0.4363298 0.4216600 0.4232301
#> [106] 0.4528662 0.4816789 0.4155394 0.4357103 0.4204575 0.4166015 0.4155394
#> [113] 0.4511944 0.4216600 0.4665677 0.4552538 0.4665677 0.4511944 0.4062517
#> [120] 0.4587111 0.4877734 0.4047713 0.4157524 0.4816789 0.4216600 0.4324049
#> [127] 0.4567567 0.4141876 0.4587111 0.4242620 0.4155394 0.4352785 0.4379200
#> [134] 0.4242620 0.4333050 0.4295286 0.4301605 0.4047713 0.4624843 0.4379200
#> [141] 0.4255681 0.4635348 0.4159352 0.4630375 0.4618223 0.4232301 0.4223613
#> [148] 0.4397252 0.4379200 0.4071515 0.4184564 0.4604887 0.4223613 0.4587111
#> [155] 0.4526385 0.4786533 0.4141965 0.4552538 0.4224480 0.4267216 0.4217379
#> [162] 0.4596140 0.4159352 0.4190884 0.4624843 0.4062517 0.4232301 0.4295286
#> [169] 0.4122334 0.4203349 0.4060208 0.4649521 0.4166015 0.4062517 0.4204575
#> [176] 0.4352785 0.4567567 0.4358095 0.4567567 0.4190884 0.4618652 0.4884184
#> [183] 0.4321504 0.4389920 0.4358095 0.4141876 0.4649521 0.4217379 0.4141965
#> [190] 0.4592040 0.4141965 0.4319108 0.4354446 0.4550241 0.4231972 0.4801132
#> [197] 0.4285069 0.4373510 0.4232301 0.4231025 0.4613656 0.4159352 0.4280896
#> [204] 0.4363298 0.4419966 0.4157524 0.4524478 0.4388243 0.4350004 0.4189114
#> [211] 0.4190520 0.4254640 0.4618652 0.4047713 0.4649521 0.4618223 0.4213793
#> [218] 0.4511944 0.4363298 0.4188853 0.4377706 0.4071515 0.4286167 0.4224480
#> [225] 0.4242620 0.4379200 0.4472967 0.4190520 0.4499097 0.4356559 0.4166015
#> [232] 0.4141876 0.4293736 0.4333050 0.4314892 0.4218537 0.4280395 0.4188853
#> [239] 0.4294843 0.4314892 0.4604887 0.4359092 0.4333050 0.4217379 0.4352785
#> [246] 0.4047713 0.4204575 0.4188853 0.4159352 0.4518823 0.4231972 0.4294843
#> [253] 0.4635348 0.4397252 0.4359772 0.4604887 0.4122334 0.4305861 0.4624843
#> [260] 0.4363298 0.4621295 0.4280896 0.4780080 0.4071515 0.4354446 0.4567567
#> [267] 0.4610356 0.4189114 0.4141965 0.4423045 0.4363298 0.4141876 0.4293736
#> [274] 0.4188853 0.4416028 0.4729848 0.4766401 0.4305861 0.4349013 0.4816789
#> [281] 0.4204575 0.4578834 0.4363298 0.4635348 0.4619478 0.4183184 0.4122334
#> [288] 0.4157524 0.4189114 0.4314892 0.4358095 0.4224480 0.4359092 0.4141876
#> [295] 0.4286167 0.4188853 0.4159352 0.4350235 0.4414303 0.4873130 0.4188853
#> [302] 0.4359092 0.4635348 0.4771841 0.4314892 0.4560523 0.4521620 0.4231025
#> [309] 0.4538960 0.4537865 0.4649521 0.4141876 0.4665677 0.4299848 0.4188853
#> [316] 0.4301605 0.4209158 0.4222715 0.4157524 0.4209158 0.4216600 0.4227967
#> [323] 0.4062517 0.4062517 0.4141965 0.4141965 0.4047713 0.4060208 0.4198962
#> [330] 0.4216600 0.4062517 0.4141965 0.4060208 0.4060208 0.4232301 0.4222715
#> [337] 0.4047713 0.4141965 0.4222715 0.4275864 0.4062517 0.4227967 0.4166015
#> [344] 0.4155394 0.4060208 0.4155394 0.4232301 0.4062517 0.4166015 0.4141965
#> [351] 0.4216600 0.4060208 0.4232301 0.4060208 0.4155394 0.4047713 0.4203349
#> [358] 0.4047713 0.4232301 0.4255681 0.4352785 0.4363298 0.4254640 0.4474994
#> [365] 0.4654252 0.4604887 0.4227967 0.4047713 0.4047713 0.4184564 0.4816789
#> [372] 0.4618223 0.4217379 0.4189114 0.4204575 0.4141965 0.4314892 0.4511944
#> [379] 0.4062517 0.4216600 0.4060208 0.4665677 0.4397252 0.4619478 0.4352785
#> [386] 0.4108191 0.4060208 0.4379200 0.4294105 0.4060208 0.4294843 0.4333050
#> [393] 0.4511944 0.4481775 0.4333050 0.4203349 0.4359140 0.4355806 0.4060208
#> [400] 0.4157524 0.4314892 0.4567567 0.4624843 0.4122334 0.4777809 0.4232301
#> [407] 0.4155394 0.4047713 0.4499097 0.4282387 0.4358095 0.4396712 0.4363298
#> [414] 0.4232301 0.4301605 0.4227967 0.4419037 0.4379200 0.4397252 0.4232301
#> [421] 0.4352785 0.4619478 0.4365855 0.4618223 0.4314892 0.4511944 0.4184564
#> [428] 0.4729848 0.4267216 0.4294843 0.4242620 0.4280395 0.4587111 0.4294105
#> [435] 0.4285069 0.4357103 0.4188853 0.4352785 0.4122334 0.4524478 0.4203349
#> [442] 0.4352785 0.4294105 0.4286167 0.4122334
# }