Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4465296 0.4252033 0.4598065 0.4405672 0.4388512 0.4252033 0.4556777
#>   [8] 0.4391198 0.4784740 0.4605605 0.4190866 0.4631516 0.4237655 0.4269303
#>  [15] 0.4249453 0.4126735 0.4729052 0.4205578 0.4748793 0.4605848 0.4342611
#>  [22] 0.4946486 0.4421245 0.4355464 0.4119538 0.4390717 0.4383609 0.4269303
#>  [29] 0.4127885 0.4355464 0.4127885 0.4796475 0.4383609 0.4490439 0.4597312
#>  [36] 0.4966280 0.4486131 0.4556777 0.4619844 0.4218913 0.4519857 0.4190866
#>  [43] 0.4246006 0.4237655 0.4278638 0.4383609 0.4246006 0.4192024 0.4486131
#>  [50] 0.4556777 0.4368645 0.4395318 0.4252704 0.4597312 0.4389581 0.4127885
#>  [57] 0.4403438 0.4735971 0.4966280 0.4323754 0.4246006 0.4375306 0.4282674
#>  [64] 0.4405672 0.4561852 0.4316134 0.4405672 0.4324885 0.4722843 0.4712263
#>  [71] 0.4335123 0.4921780 0.4295415 0.4597312 0.4820087 0.4383609 0.4538472
#>  [78] 0.4233363 0.4545077 0.4323754 0.4597312 0.4462438 0.4686713 0.4729052
#>  [85] 0.4387080 0.4482605 0.4321782 0.4392424 0.4519857 0.4363483 0.4395318
#>  [92] 0.4127885 0.4383609 0.4519857 0.4593934 0.4597312 0.4183930 0.4788689
#>  [99] 0.4245911 0.4329666 0.4218913 0.4255253 0.4383609 0.4252704 0.4260136
#> [106] 0.4530342 0.4729052 0.4192024 0.4389581 0.4255253 0.4207303 0.4192024
#> [113] 0.4490439 0.4252704 0.4617784 0.4519857 0.4617784 0.4490439 0.4126735
#> [120] 0.4556777 0.4790197 0.4119538 0.4190866 0.4729052 0.4252704 0.4335123
#> [127] 0.4538472 0.4205578 0.4556777 0.4278638 0.4192024 0.4384319 0.4392424
#> [134] 0.4278638 0.4355464 0.4329666 0.4321782 0.4119538 0.4601045 0.4392424
#> [141] 0.4283838 0.4597312 0.4218913 0.4619844 0.4602790 0.4260136 0.4243953
#> [148] 0.4405672 0.4392424 0.4143498 0.4249324 0.4569324 0.4243953 0.4556777
#> [155] 0.4525475 0.4709234 0.4183930 0.4519857 0.4269303 0.4285434 0.4252033
#> [162] 0.4568260 0.4218913 0.4255951 0.4601045 0.4126735 0.4260136 0.4329666
#> [169] 0.4187665 0.4233363 0.4127885 0.4605848 0.4207303 0.4126735 0.4255253
#> [176] 0.4384319 0.4538472 0.4399397 0.4538472 0.4255951 0.4605605 0.4787426
#> [183] 0.4352514 0.4406544 0.4399397 0.4205578 0.4605848 0.4252033 0.4183930
#> [190] 0.4607252 0.4183930 0.4335774 0.4397579 0.4521099 0.4301153 0.4709458
#> [197] 0.4316134 0.4390750 0.4260136 0.4259347 0.4603603 0.4218913 0.4303682
#> [204] 0.4383609 0.4439939 0.4190866 0.4534975 0.4387080 0.4363483 0.4261295
#> [211] 0.4265576 0.4288821 0.4605605 0.4119538 0.4605848 0.4602790 0.4282674
#> [218] 0.4490439 0.4383609 0.4245911 0.4414384 0.4143498 0.4311518 0.4269303
#> [225] 0.4278638 0.4392424 0.4474143 0.4265576 0.4471055 0.4378933 0.4207303
#> [232] 0.4205578 0.4351122 0.4355464 0.4342611 0.4263136 0.4306002 0.4245911
#> [239] 0.4323754 0.4342611 0.4569324 0.4390717 0.4355464 0.4252033 0.4384319
#> [246] 0.4119538 0.4255253 0.4245911 0.4218913 0.4495178 0.4301153 0.4323754
#> [253] 0.4597312 0.4405672 0.4375306 0.4569324 0.4187665 0.4338770 0.4601045
#> [260] 0.4383609 0.4598065 0.4303682 0.4793656 0.4143498 0.4397579 0.4538472
#> [267] 0.4598145 0.4261295 0.4183930 0.4496951 0.4383609 0.4205578 0.4351122
#> [274] 0.4245911 0.4456183 0.4704247 0.4712030 0.4338770 0.4391198 0.4729052
#> [281] 0.4255253 0.4593934 0.4383609 0.4597312 0.4595747 0.4257655 0.4187665
#> [288] 0.4190866 0.4261295 0.4342611 0.4399397 0.4269303 0.4390717 0.4205578
#> [295] 0.4311518 0.4245911 0.4218913 0.4368645 0.4426396 0.4786573 0.4245911
#> [302] 0.4390717 0.4597312 0.4707331 0.4342611 0.4629445 0.4513941 0.4259347
#> [309] 0.4548382 0.4549583 0.4605848 0.4205578 0.4617784 0.4335647 0.4245911
#> [316] 0.4321782 0.4237655 0.4246006 0.4190866 0.4237655 0.4252704 0.4260204
#> [323] 0.4126735 0.4126735 0.4183930 0.4183930 0.4119538 0.4127885 0.4241738
#> [330] 0.4252704 0.4126735 0.4183930 0.4127885 0.4127885 0.4260136 0.4246006
#> [337] 0.4119538 0.4183930 0.4246006 0.4302136 0.4126735 0.4260204 0.4207303
#> [344] 0.4192024 0.4127885 0.4192024 0.4260136 0.4126735 0.4207303 0.4183930
#> [351] 0.4252704 0.4127885 0.4260136 0.4127885 0.4192024 0.4119538 0.4233363
#> [358] 0.4119538 0.4260136 0.4283838 0.4384319 0.4383609 0.4288821 0.4500850
#> [365] 0.4636483 0.4569324 0.4260204 0.4119538 0.4119538 0.4249324 0.4729052
#> [372] 0.4602790 0.4252033 0.4261295 0.4255253 0.4183930 0.4342611 0.4490439
#> [379] 0.4126735 0.4252704 0.4127885 0.4617784 0.4405672 0.4595747 0.4384319
#> [386] 0.4169842 0.4127885 0.4392424 0.4323200 0.4127885 0.4323754 0.4355464
#> [393] 0.4490439 0.4462438 0.4355464 0.4233363 0.4375742 0.4395318 0.4127885
#> [400] 0.4190866 0.4342611 0.4538472 0.4601045 0.4187665 0.4712263 0.4260136
#> [407] 0.4192024 0.4119538 0.4471055 0.4316563 0.4399397 0.4432344 0.4383609
#> [414] 0.4260136 0.4321782 0.4260204 0.4460515 0.4392424 0.4405672 0.4260136
#> [421] 0.4384319 0.4595747 0.4372039 0.4602790 0.4342611 0.4490439 0.4249324
#> [428] 0.4704247 0.4285434 0.4323754 0.4278638 0.4306002 0.4556777 0.4323200
#> [435] 0.4316134 0.4389581 0.4245911 0.4384319 0.4187665 0.4534975 0.4233363
#> [442] 0.4384319 0.4323200 0.4311518 0.4187665