Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4401765 0.4167916 0.4582170 0.4344503 0.4318485 0.4167916 0.4546557
#>   [8] 0.4306249 0.4862068 0.4587142 0.4103478 0.4663723 0.4153475 0.4174171
#>  [15] 0.4172004 0.4008666 0.4779777 0.4092733 0.4760221 0.4601851 0.4262893
#>  [22] 0.5035692 0.4354621 0.4282673 0.3996242 0.4309379 0.4306850 0.4174171
#>  [29] 0.4011878 0.4282673 0.4011878 0.4789951 0.4306850 0.4472984 0.4587944
#>  [36] 0.5055378 0.4448616 0.4546557 0.4573754 0.4111983 0.4506831 0.4103478
#>  [43] 0.4170223 0.4153475 0.4191903 0.4306850 0.4170223 0.4107222 0.4448616
#>  [50] 0.4546557 0.4298579 0.4310060 0.4160031 0.4587944 0.4308629 0.4011878
#>  [57] 0.4341270 0.4753458 0.5055378 0.4242928 0.4170223 0.4304876 0.4172925
#>  [64] 0.4344503 0.4543382 0.4236192 0.4344503 0.4246436 0.4711613 0.4740456
#>  [71] 0.4272410 0.4968807 0.4216066 0.4587944 0.4893648 0.4306850 0.4526511
#>  [78] 0.4150373 0.4492365 0.4242928 0.4587944 0.4437734 0.4684929 0.4779777
#>  [85] 0.4334029 0.4414956 0.4248503 0.4322280 0.4506831 0.4292660 0.4310060
#>  [92] 0.4011878 0.4306850 0.4506831 0.4538766 0.4587944 0.4090717 0.4806394
#>  [99] 0.4135712 0.4246742 0.4111983 0.4152267 0.4306850 0.4160031 0.4173892
#> [106] 0.4481114 0.4779777 0.4107222 0.4308629 0.4152267 0.4119919 0.4107222
#> [113] 0.4472984 0.4160031 0.4623149 0.4506831 0.4623149 0.4472984 0.4008666
#> [120] 0.4546557 0.4842995 0.3996242 0.4103478 0.4779777 0.4160031 0.4272410
#> [127] 0.4526511 0.4092733 0.4546557 0.4191903 0.4107222 0.4305045 0.4322280
#> [134] 0.4191903 0.4282673 0.4246742 0.4248503 0.3996242 0.4584374 0.4322280
#> [141] 0.4207684 0.4587944 0.4111983 0.4573754 0.4582543 0.4173892 0.4165948
#> [148] 0.4344503 0.4322280 0.4025231 0.4137505 0.4563777 0.4165948 0.4546557
#> [155] 0.4481829 0.4746061 0.4090717 0.4506831 0.4174171 0.4211625 0.4167916
#> [162] 0.4550082 0.4111983 0.4141967 0.4584374 0.4008666 0.4173892 0.4246742
#> [169] 0.4073468 0.4150373 0.4011878 0.4601851 0.4119919 0.4008666 0.4152267
#> [176] 0.4305045 0.4526511 0.4315885 0.4526511 0.4141967 0.4587142 0.4845111
#> [183] 0.4276105 0.4341397 0.4315885 0.4092733 0.4601851 0.4167916 0.4090717
#> [190] 0.4543902 0.4090717 0.4267801 0.4312069 0.4499552 0.4190099 0.4762237
#> [197] 0.4236192 0.4332215 0.4173892 0.4181381 0.4582620 0.4111983 0.4228739
#> [204] 0.4306850 0.4374869 0.4103478 0.4481780 0.4334029 0.4292660 0.4143951
#> [211] 0.4148674 0.4210383 0.4587142 0.3996242 0.4601851 0.4582543 0.4172925
#> [218] 0.4472984 0.4306850 0.4135712 0.4337344 0.4025231 0.4232276 0.4174171
#> [225] 0.4191903 0.4322280 0.4419664 0.4148674 0.4454359 0.4307284 0.4119919
#> [232] 0.4092733 0.4251161 0.4282673 0.4262893 0.4180742 0.4223714 0.4135712
#> [239] 0.4242928 0.4262893 0.4563777 0.4309379 0.4282673 0.4167916 0.4305045
#> [246] 0.3996242 0.4152267 0.4135712 0.4111983 0.4465521 0.4190099 0.4242928
#> [253] 0.4587944 0.4344503 0.4304876 0.4563777 0.4073468 0.4266006 0.4584374
#> [260] 0.4306850 0.4582170 0.4228739 0.4742878 0.4025231 0.4312069 0.4526511
#> [267] 0.4578571 0.4143951 0.4090717 0.4381351 0.4306850 0.4092733 0.4251161
#> [274] 0.4135712 0.4377716 0.4703090 0.4737529 0.4266006 0.4306249 0.4779777
#> [281] 0.4152267 0.4538766 0.4306850 0.4587944 0.4580813 0.4140783 0.4073468
#> [288] 0.4103478 0.4143951 0.4262893 0.4315885 0.4174171 0.4309379 0.4092733
#> [295] 0.4232276 0.4135712 0.4111983 0.4298579 0.4367968 0.4840225 0.4135712
#> [302] 0.4309379 0.4587944 0.4737591 0.4262893 0.4515603 0.4477560 0.4181381
#> [309] 0.4488976 0.4502803 0.4601851 0.4092733 0.4623149 0.4253200 0.4135712
#> [316] 0.4248503 0.4153475 0.4170223 0.4103478 0.4153475 0.4160031 0.4175062
#> [323] 0.4008666 0.4008666 0.4090717 0.4090717 0.3996242 0.4011878 0.4154766
#> [330] 0.4160031 0.4008666 0.4090717 0.4011878 0.4011878 0.4173892 0.4170223
#> [337] 0.3996242 0.4090717 0.4170223 0.4227290 0.4008666 0.4175062 0.4119919
#> [344] 0.4107222 0.4011878 0.4107222 0.4173892 0.4008666 0.4119919 0.4090717
#> [351] 0.4160031 0.4011878 0.4173892 0.4011878 0.4107222 0.3996242 0.4150373
#> [358] 0.3996242 0.4173892 0.4207684 0.4305045 0.4306850 0.4210383 0.4434634
#> [365] 0.4623518 0.4563777 0.4175062 0.3996242 0.3996242 0.4137505 0.4779777
#> [372] 0.4582543 0.4167916 0.4143951 0.4152267 0.4090717 0.4262893 0.4472984
#> [379] 0.4008666 0.4160031 0.4011878 0.4623149 0.4344503 0.4580813 0.4305045
#> [386] 0.4055635 0.4011878 0.4322280 0.4242570 0.4011878 0.4242928 0.4282673
#> [393] 0.4472984 0.4437734 0.4282673 0.4150373 0.4306526 0.4310060 0.4011878
#> [400] 0.4103478 0.4262893 0.4526511 0.4584374 0.4073468 0.4740456 0.4173892
#> [407] 0.4107222 0.3996242 0.4454359 0.4242009 0.4315885 0.4353324 0.4306850
#> [414] 0.4173892 0.4248503 0.4175062 0.4378055 0.4322280 0.4344503 0.4173892
#> [421] 0.4305045 0.4580813 0.4309713 0.4582543 0.4262893 0.4472984 0.4137505
#> [428] 0.4703090 0.4211625 0.4242928 0.4191903 0.4223714 0.4546557 0.4242570
#> [435] 0.4236192 0.4308629 0.4135712 0.4305045 0.4073468 0.4481780 0.4150373
#> [442] 0.4305045 0.4242570 0.4232276 0.4073468