Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4399281 0.4141823 0.4606127 0.4336291 0.4301992 0.4141823 0.4553133
#>   [8] 0.4304371 0.4864426 0.4602204 0.4068262 0.4669281 0.4120097 0.4154690
#>  [15] 0.4123289 0.3971109 0.4798470 0.4060716 0.4786292 0.4619875 0.4243438
#>  [22] 0.5069748 0.4353566 0.4267806 0.3958438 0.4308999 0.4301364 0.4154690
#>  [29] 0.3963356 0.4267806 0.3963356 0.4809431 0.4301364 0.4474006 0.4608114
#>  [36] 0.5091603 0.4432766 0.4553133 0.4594554 0.4085475 0.4512698 0.4068262
#>  [43] 0.4125275 0.4120097 0.4168702 0.4301364 0.4125275 0.4059646 0.4432766
#>  [50] 0.4553133 0.4286799 0.4309896 0.4135423 0.4608114 0.4309145 0.3963356
#>  [57] 0.4331486 0.4754017 0.5091603 0.4227382 0.4125275 0.4296615 0.4157879
#>  [64] 0.4336291 0.4539545 0.4223199 0.4336291 0.4215279 0.4714993 0.4766379
#>  [71] 0.4254149 0.5022760 0.4182203 0.4608114 0.4904954 0.4301364 0.4537066
#>  [78] 0.4119592 0.4497232 0.4227382 0.4608114 0.4436228 0.4712050 0.4798470
#>  [85] 0.4317270 0.4421367 0.4220274 0.4315359 0.4512698 0.4278860 0.4309896
#>  [92] 0.3963356 0.4301364 0.4512698 0.4551284 0.4608114 0.4055014 0.4837845
#>  [99] 0.4118242 0.4235223 0.4085475 0.4132739 0.4301364 0.4135423 0.4148715
#> [106] 0.4488874 0.4798470 0.4059646 0.4309145 0.4132739 0.4086240 0.4059646
#> [113] 0.4474006 0.4135423 0.4638854 0.4512698 0.4638854 0.4474006 0.3971109
#> [120] 0.4553133 0.4869773 0.3958438 0.4068262 0.4798470 0.4135423 0.4254149
#> [127] 0.4537066 0.4060716 0.4553133 0.4168702 0.4059646 0.4302199 0.4315359
#> [134] 0.4168702 0.4267806 0.4235223 0.4220274 0.3958438 0.4607657 0.4315359
#> [141] 0.4182407 0.4608114 0.4085475 0.4594554 0.4606222 0.4148715 0.4131474
#> [148] 0.4336291 0.4315359 0.3989427 0.4122920 0.4575666 0.4131474 0.4553133
#> [155] 0.4484077 0.4766724 0.4055014 0.4512698 0.4154690 0.4182268 0.4141823
#> [162] 0.4544099 0.4085475 0.4129936 0.4607657 0.3971109 0.4148715 0.4235223
#> [169] 0.4045311 0.4119592 0.3963356 0.4619875 0.4086240 0.3971109 0.4132739
#> [176] 0.4302199 0.4537066 0.4307113 0.4537066 0.4129936 0.4602204 0.4869105
#> [183] 0.4263731 0.4329751 0.4307113 0.4060716 0.4619875 0.4141823 0.4055014
#> [190] 0.4568114 0.4055014 0.4245251 0.4308827 0.4483653 0.4178055 0.4771490
#> [197] 0.4223199 0.4312425 0.4148715 0.4149273 0.4603007 0.4085475 0.4204788
#> [204] 0.4301364 0.4375726 0.4068262 0.4485622 0.4317270 0.4278860 0.4132419
#> [211] 0.4134106 0.4178376 0.4602204 0.3958438 0.4619875 0.4606222 0.4157879
#> [218] 0.4474006 0.4301364 0.4118242 0.4332514 0.3989427 0.4216101 0.4154690
#> [225] 0.4168702 0.4315359 0.4418719 0.4134106 0.4452453 0.4296983 0.4086240
#> [232] 0.4060716 0.4246654 0.4267806 0.4243438 0.4145765 0.4204895 0.4118242
#> [239] 0.4227382 0.4243438 0.4575666 0.4308999 0.4267806 0.4141823 0.4302199
#> [246] 0.3958438 0.4132739 0.4118242 0.4085475 0.4446700 0.4178055 0.4227382
#> [253] 0.4608114 0.4336291 0.4296615 0.4575666 0.4045311 0.4240430 0.4607657
#> [260] 0.4301364 0.4606127 0.4204788 0.4758135 0.3989427 0.4308827 0.4537066
#> [267] 0.4600151 0.4132419 0.4055014 0.4400478 0.4301364 0.4060716 0.4246654
#> [274] 0.4118242 0.4378284 0.4733478 0.4760463 0.4240430 0.4304371 0.4798470
#> [281] 0.4132739 0.4551284 0.4301364 0.4608114 0.4602655 0.4127859 0.4045311
#> [288] 0.4068262 0.4132419 0.4243438 0.4307113 0.4154690 0.4308999 0.4060716
#> [295] 0.4216101 0.4118242 0.4085475 0.4286799 0.4351586 0.4867723 0.4118242
#> [302] 0.4308999 0.4608114 0.4762020 0.4243438 0.4547567 0.4455865 0.4149273
#> [309] 0.4505334 0.4495359 0.4619875 0.4060716 0.4638854 0.4235682 0.4118242
#> [316] 0.4220274 0.4120097 0.4125275 0.4068262 0.4120097 0.4135423 0.4138792
#> [323] 0.3971109 0.3971109 0.4055014 0.4055014 0.3958438 0.3963356 0.4117725
#> [330] 0.4135423 0.3971109 0.4055014 0.3963356 0.3963356 0.4148715 0.4125275
#> [337] 0.3958438 0.4055014 0.4125275 0.4191296 0.3971109 0.4138792 0.4086240
#> [344] 0.4059646 0.3963356 0.4059646 0.4148715 0.3971109 0.4086240 0.4055014
#> [351] 0.4135423 0.3963356 0.4148715 0.3963356 0.4059646 0.3958438 0.4119592
#> [358] 0.3958438 0.4148715 0.4182407 0.4302199 0.4301364 0.4178376 0.4443174
#> [365] 0.4644620 0.4575666 0.4138792 0.3958438 0.3958438 0.4122920 0.4798470
#> [372] 0.4606222 0.4141823 0.4132419 0.4132739 0.4055014 0.4243438 0.4474006
#> [379] 0.3971109 0.4135423 0.3963356 0.4638854 0.4336291 0.4602655 0.4302199
#> [386] 0.4023203 0.3963356 0.4315359 0.4231366 0.3963356 0.4227382 0.4267806
#> [393] 0.4474006 0.4436228 0.4267806 0.4119592 0.4297478 0.4309896 0.3963356
#> [400] 0.4068262 0.4243438 0.4537066 0.4607657 0.4045311 0.4766379 0.4148715
#> [407] 0.4059646 0.3958438 0.4452453 0.4209102 0.4307113 0.4352738 0.4301364
#> [414] 0.4148715 0.4220274 0.4138792 0.4383639 0.4315359 0.4336291 0.4148715
#> [421] 0.4302199 0.4602655 0.4293643 0.4606222 0.4243438 0.4474006 0.4122920
#> [428] 0.4733478 0.4182268 0.4227382 0.4168702 0.4204895 0.4553133 0.4231366
#> [435] 0.4223199 0.4309145 0.4118242 0.4302199 0.4045311 0.4485622 0.4119592
#> [442] 0.4302199 0.4231366 0.4216101 0.4045311