Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4477243 0.4267749 0.4615106 0.4413129 0.4399079 0.4267749 0.4567655
#>   [8] 0.4400908 0.4800246 0.4620751 0.4209303 0.4639835 0.4245035 0.4276626
#>  [15] 0.4262223 0.4140032 0.4739122 0.4208295 0.4758438 0.4614440 0.4345220
#>  [22] 0.4954751 0.4435169 0.4358859 0.4128300 0.4395214 0.4381875 0.4276626
#>  [29] 0.4135583 0.4358859 0.4135583 0.4798390 0.4381875 0.4505815 0.4602585
#>  [36] 0.4970513 0.4491565 0.4567655 0.4617847 0.4221998 0.4533475 0.4209303
#>  [43] 0.4252273 0.4245035 0.4285482 0.4381875 0.4252273 0.4204602 0.4491565
#>  [50] 0.4567655 0.4377027 0.4403843 0.4260575 0.4602585 0.4399840 0.4135583
#>  [57] 0.4423014 0.4747471 0.4970513 0.4334995 0.4252273 0.4381441 0.4294449
#>  [64] 0.4413129 0.4572513 0.4329715 0.4413129 0.4332314 0.4725873 0.4719635
#>  [71] 0.4348899 0.4920573 0.4300189 0.4602585 0.4832694 0.4381875 0.4557466
#>  [78] 0.4245128 0.4550227 0.4334995 0.4602585 0.4478540 0.4698737 0.4739122
#>  [85] 0.4395466 0.4489265 0.4323831 0.4394765 0.4533475 0.4361443 0.4403843
#>  [92] 0.4135583 0.4381875 0.4533475 0.4594368 0.4602585 0.4197327 0.4795639
#>  [99] 0.4244217 0.4344823 0.4221998 0.4257394 0.4381875 0.4260575 0.4273977
#> [106] 0.4542413 0.4739122 0.4204602 0.4399840 0.4257394 0.4219472 0.4204602
#> [113] 0.4505815 0.4260575 0.4632029 0.4533475 0.4632029 0.4505815 0.4140032
#> [120] 0.4567655 0.4795274 0.4128300 0.4209303 0.4739122 0.4260575 0.4348899
#> [127] 0.4557466 0.4208295 0.4567655 0.4285482 0.4204602 0.4391636 0.4394765
#> [134] 0.4285482 0.4358859 0.4344823 0.4323831 0.4128300 0.4612357 0.4394765
#> [141] 0.4291672 0.4602585 0.4221998 0.4617847 0.4618179 0.4273977 0.4256421
#> [148] 0.4413129 0.4394765 0.4150735 0.4257205 0.4579943 0.4256421 0.4567655
#> [155] 0.4536980 0.4709690 0.4197327 0.4533475 0.4276626 0.4291724 0.4267749
#> [162] 0.4585200 0.4221998 0.4261064 0.4612357 0.4140032 0.4273977 0.4344823
#> [169] 0.4198546 0.4245128 0.4135583 0.4614440 0.4219472 0.4140032 0.4257394
#> [176] 0.4391636 0.4557466 0.4407357 0.4557466 0.4261064 0.4620751 0.4784004
#> [183] 0.4370546 0.4419809 0.4407357 0.4208295 0.4614440 0.4267749 0.4197327
#> [190] 0.4601551 0.4197327 0.4338208 0.4406905 0.4527493 0.4312273 0.4714742
#> [197] 0.4329715 0.4403682 0.4273977 0.4265956 0.4619433 0.4221998 0.4314131
#> [204] 0.4381875 0.4455567 0.4209303 0.4542681 0.4395466 0.4361443 0.4270553
#> [211] 0.4276061 0.4299307 0.4620751 0.4128300 0.4614440 0.4618179 0.4294449
#> [218] 0.4505815 0.4381875 0.4244217 0.4425352 0.4150735 0.4314638 0.4276626
#> [225] 0.4285482 0.4394765 0.4484500 0.4276061 0.4493493 0.4391120 0.4219472
#> [232] 0.4208295 0.4359021 0.4358859 0.4345220 0.4273581 0.4312288 0.4244217
#> [239] 0.4334995 0.4345220 0.4579943 0.4395214 0.4358859 0.4267749 0.4391636
#> [246] 0.4128300 0.4257394 0.4244217 0.4221998 0.4504632 0.4312273 0.4334995
#> [253] 0.4602585 0.4413129 0.4381441 0.4579943 0.4198546 0.4346406 0.4612357
#> [260] 0.4381875 0.4615106 0.4314131 0.4797532 0.4150735 0.4406905 0.4557466
#> [267] 0.4614987 0.4270553 0.4197327 0.4501224 0.4381875 0.4208295 0.4359021
#> [274] 0.4244217 0.4461234 0.4718956 0.4725786 0.4346406 0.4400908 0.4739122
#> [281] 0.4257394 0.4594368 0.4381875 0.4602585 0.4609171 0.4268268 0.4198546
#> [288] 0.4209303 0.4270553 0.4345220 0.4407357 0.4276626 0.4395214 0.4208295
#> [295] 0.4314638 0.4244217 0.4221998 0.4377027 0.4428625 0.4797032 0.4244217
#> [302] 0.4395214 0.4602585 0.4716144 0.4345220 0.4617329 0.4531187 0.4265956
#> [309] 0.4550445 0.4555156 0.4614440 0.4208295 0.4632029 0.4350419 0.4244217
#> [316] 0.4323831 0.4245035 0.4252273 0.4209303 0.4245035 0.4260575 0.4267597
#> [323] 0.4140032 0.4140032 0.4197327 0.4197327 0.4128300 0.4135583 0.4253783
#> [330] 0.4260575 0.4140032 0.4197327 0.4135583 0.4135583 0.4273977 0.4252273
#> [337] 0.4128300 0.4197327 0.4252273 0.4310286 0.4140032 0.4267597 0.4219472
#> [344] 0.4204602 0.4135583 0.4204602 0.4273977 0.4140032 0.4219472 0.4197327
#> [351] 0.4260575 0.4135583 0.4273977 0.4135583 0.4204602 0.4128300 0.4245128
#> [358] 0.4128300 0.4273977 0.4291672 0.4391636 0.4381875 0.4299307 0.4509698
#> [365] 0.4652770 0.4579943 0.4267597 0.4128300 0.4128300 0.4257205 0.4739122
#> [372] 0.4618179 0.4267749 0.4270553 0.4257394 0.4197327 0.4345220 0.4505815
#> [379] 0.4140032 0.4260575 0.4135583 0.4632029 0.4413129 0.4609171 0.4391636
#> [386] 0.4176216 0.4135583 0.4394765 0.4333699 0.4135583 0.4334995 0.4358859
#> [393] 0.4505815 0.4478540 0.4358859 0.4245128 0.4386973 0.4403843 0.4135583
#> [400] 0.4209303 0.4345220 0.4557466 0.4612357 0.4198546 0.4719635 0.4273977
#> [407] 0.4204602 0.4128300 0.4493493 0.4319552 0.4407357 0.4442317 0.4381875
#> [414] 0.4273977 0.4323831 0.4267597 0.4466949 0.4394765 0.4413129 0.4273977
#> [421] 0.4391636 0.4609171 0.4375033 0.4618179 0.4345220 0.4505815 0.4257205
#> [428] 0.4718956 0.4291724 0.4334995 0.4285482 0.4312288 0.4567655 0.4333699
#> [435] 0.4329715 0.4399840 0.4244217 0.4391636 0.4198546 0.4542681 0.4245128
#> [442] 0.4391636 0.4333699 0.4314638 0.4198546
# }