Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4395362 0.4132637 0.4571108 0.4325441 0.4305584 0.4132637 0.4520958
#>   [8] 0.4301788 0.4815421 0.4580660 0.4060571 0.4629551 0.4114241 0.4142425
#>  [15] 0.4135642 0.3972091 0.4747449 0.4060261 0.4750721 0.4581085 0.4244748
#>  [22] 0.5009203 0.4333749 0.4261958 0.3957039 0.4300308 0.4294957 0.4142425
#>  [29] 0.3971793 0.4261958 0.3971793 0.4789334 0.4294957 0.4439389 0.4567944
#>  [36] 0.5030658 0.4429144 0.4520958 0.4578234 0.4077508 0.4478773 0.4060571
#>  [43] 0.4130064 0.4114241 0.4165150 0.4294957 0.4130064 0.4061872 0.4429144
#>  [50] 0.4520958 0.4273967 0.4305601 0.4134223 0.4567944 0.4304013 0.3971793
#>  [57] 0.4317980 0.4721179 0.5030658 0.4226851 0.4130064 0.4281119 0.4155242
#>  [64] 0.4325441 0.4508712 0.4207507 0.4325441 0.4220350 0.4706737 0.4720241
#>  [71] 0.4236794 0.4976168 0.4181631 0.4567944 0.4852111 0.4294957 0.4503148
#>  [78] 0.4109304 0.4492171 0.4226851 0.4567944 0.4405375 0.4671452 0.4747449
#>  [85] 0.4300998 0.4411501 0.4215440 0.4309020 0.4478773 0.4266307 0.4305601
#>  [92] 0.3971793 0.4294957 0.4478773 0.4546029 0.4567944 0.4046219 0.4795792
#>  [99] 0.4109602 0.4224782 0.4077508 0.4124613 0.4294957 0.4134223 0.4150246
#> [106] 0.4468024 0.4747449 0.4061872 0.4304013 0.4124613 0.4073812 0.4061872
#> [113] 0.4439389 0.4134223 0.4595944 0.4478773 0.4595944 0.4439389 0.3972091
#> [120] 0.4520958 0.4814636 0.3957039 0.4060571 0.4747449 0.4134223 0.4236794
#> [127] 0.4503148 0.4060261 0.4520958 0.4165150 0.4061872 0.4293817 0.4309020
#> [134] 0.4165150 0.4261958 0.4224782 0.4215440 0.3957039 0.4568653 0.4309020
#> [141] 0.4168474 0.4567944 0.4077508 0.4578234 0.4570909 0.4150246 0.4127945
#> [148] 0.4325441 0.4309020 0.3985591 0.4112207 0.4536237 0.4127945 0.4520958
#> [155] 0.4461708 0.4720278 0.4046219 0.4478773 0.4142425 0.4174547 0.4132637
#> [162] 0.4518872 0.4077508 0.4118821 0.4568653 0.3972091 0.4150246 0.4224782
#> [169] 0.4042821 0.4109304 0.3971793 0.4581085 0.4073812 0.3972091 0.4124613
#> [176] 0.4293817 0.4503148 0.4317332 0.4503148 0.4118821 0.4580660 0.4813551
#> [183] 0.4254943 0.4323489 0.4317332 0.4060261 0.4581085 0.4132637 0.4046219
#> [190] 0.4556264 0.4046219 0.4233046 0.4315353 0.4460812 0.4174257 0.4722630
#> [197] 0.4207507 0.4305774 0.4150246 0.4140207 0.4577694 0.4077508 0.4197382
#> [204] 0.4294957 0.4356305 0.4060571 0.4470794 0.4300998 0.4266307 0.4126376
#> [211] 0.4136973 0.4179553 0.4580660 0.3957039 0.4581085 0.4570909 0.4155242
#> [218] 0.4439389 0.4294957 0.4109602 0.4332935 0.3985591 0.4201699 0.4142425
#> [225] 0.4165150 0.4309020 0.4401513 0.4136973 0.4423476 0.4290146 0.4073812
#> [232] 0.4060261 0.4234184 0.4261958 0.4244748 0.4146097 0.4202911 0.4109602
#> [239] 0.4226851 0.4244748 0.4536237 0.4300308 0.4261958 0.4132637 0.4293817
#> [246] 0.3957039 0.4124613 0.4109602 0.4077508 0.4443140 0.4174257 0.4226851
#> [253] 0.4567944 0.4325441 0.4281119 0.4536237 0.4042821 0.4238003 0.4568653
#> [260] 0.4294957 0.4571108 0.4197382 0.4755737 0.3985591 0.4315353 0.4503148
#> [267] 0.4565866 0.4126376 0.4046219 0.4392099 0.4294957 0.4060261 0.4234184
#> [274] 0.4109602 0.4362405 0.4693406 0.4728667 0.4238003 0.4301788 0.4747449
#> [281] 0.4124613 0.4546029 0.4294957 0.4567944 0.4562639 0.4123306 0.4042821
#> [288] 0.4060571 0.4126376 0.4244748 0.4317332 0.4142425 0.4300308 0.4060261
#> [295] 0.4201699 0.4109602 0.4077508 0.4273967 0.4346932 0.4815549 0.4109602
#> [302] 0.4300308 0.4567944 0.4714500 0.4244748 0.4553177 0.4465836 0.4140207
#> [309] 0.4482115 0.4487833 0.4581085 0.4060261 0.4595944 0.4237542 0.4109602
#> [316] 0.4215440 0.4114241 0.4130064 0.4060571 0.4114241 0.4134223 0.4147903
#> [323] 0.3972091 0.3972091 0.4046219 0.4046219 0.3957039 0.3971793 0.4120302
#> [330] 0.4134223 0.3972091 0.4046219 0.3971793 0.3971793 0.4150246 0.4130064
#> [337] 0.3957039 0.4046219 0.4130064 0.4194575 0.3972091 0.4147903 0.4073812
#> [344] 0.4061872 0.3971793 0.4061872 0.4150246 0.3972091 0.4073812 0.4046219
#> [351] 0.4134223 0.3971793 0.4150246 0.3971793 0.4061872 0.3957039 0.4109304
#> [358] 0.3957039 0.4150246 0.4168474 0.4293817 0.4294957 0.4179553 0.4434518
#> [365] 0.4613054 0.4536237 0.4147903 0.3957039 0.3957039 0.4112207 0.4747449
#> [372] 0.4570909 0.4132637 0.4126376 0.4124613 0.4046219 0.4244748 0.4439389
#> [379] 0.3972091 0.4134223 0.3971793 0.4595944 0.4325441 0.4562639 0.4293817
#> [386] 0.4020695 0.3971793 0.4309020 0.4214496 0.3971793 0.4226851 0.4261958
#> [393] 0.4439389 0.4405375 0.4261958 0.4109304 0.4287207 0.4305601 0.3971793
#> [400] 0.4060571 0.4244748 0.4503148 0.4568653 0.4042821 0.4720241 0.4150246
#> [407] 0.4061872 0.3957039 0.4423476 0.4206453 0.4317332 0.4351543 0.4294957
#> [414] 0.4150246 0.4215440 0.4147903 0.4368808 0.4309020 0.4325441 0.4150246
#> [421] 0.4293817 0.4562639 0.4281783 0.4570909 0.4244748 0.4439389 0.4112207
#> [428] 0.4693406 0.4174547 0.4226851 0.4165150 0.4202911 0.4520958 0.4214496
#> [435] 0.4207507 0.4304013 0.4109602 0.4293817 0.4042821 0.4470794 0.4109304
#> [442] 0.4293817 0.4214496 0.4201699 0.4042821