Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4419433 0.4180154 0.4588127 0.4357936 0.4336880 0.4180154 0.4537749
#>   [8] 0.4335942 0.4814600 0.4590372 0.4120341 0.4635742 0.4177106 0.4198039
#>  [15] 0.4192499 0.4042617 0.4751840 0.4118355 0.4749727 0.4598453 0.4277836
#>  [22] 0.4981025 0.4366001 0.4295881 0.4034661 0.4334536 0.4326274 0.4198039
#>  [29] 0.4044682 0.4295881 0.4044682 0.4785184 0.4326274 0.4469110 0.4585301
#>  [36] 0.5001110 0.4453851 0.4537749 0.4582966 0.4136277 0.4497658 0.4120341
#>  [43] 0.4187936 0.4177106 0.4211385 0.4326274 0.4187936 0.4123313 0.4453851
#>  [50] 0.4537749 0.4318621 0.4336368 0.4190151 0.4585301 0.4336149 0.4044682
#>  [57] 0.4351815 0.4734441 0.5001110 0.4262081 0.4187936 0.4323469 0.4212564
#>  [64] 0.4357936 0.4542519 0.4254027 0.4357936 0.4261704 0.4709298 0.4726047
#>  [71] 0.4279061 0.4949956 0.4239668 0.4585301 0.4855685 0.4326274 0.4521496
#>  [78] 0.4157144 0.4505222 0.4262081 0.4585301 0.4444685 0.4678710 0.4751840
#>  [85] 0.4342210 0.4433081 0.4260299 0.4340273 0.4497658 0.4308897 0.4336368
#>  [92] 0.4044682 0.4326274 0.4497658 0.4555278 0.4585301 0.4112491 0.4790317
#>  [99] 0.4165906 0.4264292 0.4136277 0.4179981 0.4326274 0.4190151 0.4199451
#> [106] 0.4486444 0.4751840 0.4123313 0.4336149 0.4179981 0.4130683 0.4123313
#> [113] 0.4469110 0.4190151 0.4613698 0.4497658 0.4613698 0.4469110 0.4042617
#> [120] 0.4537749 0.4810319 0.4034661 0.4120341 0.4751840 0.4190151 0.4279061
#> [127] 0.4521496 0.4118355 0.4537749 0.4211385 0.4123313 0.4330907 0.4340273
#> [134] 0.4211385 0.4295881 0.4264292 0.4260299 0.4034661 0.4588032 0.4340273
#> [141] 0.4214717 0.4585301 0.4136277 0.4582966 0.4586805 0.4199451 0.4184291
#> [148] 0.4357936 0.4340273 0.4053215 0.4172327 0.4555444 0.4184291 0.4537749
#> [155] 0.4479577 0.4723828 0.4112491 0.4497658 0.4198039 0.4220477 0.4180154
#> [162] 0.4549680 0.4136277 0.4176506 0.4588032 0.4042617 0.4199451 0.4264292
#> [169] 0.4102056 0.4157144 0.4044682 0.4598453 0.4130683 0.4042617 0.4179981
#> [176] 0.4330907 0.4521496 0.4342025 0.4521496 0.4176506 0.4590372 0.4808289
#> [183] 0.4296905 0.4361651 0.4342025 0.4118355 0.4598453 0.4180154 0.4112491
#> [190] 0.4565655 0.4112491 0.4278162 0.4346891 0.4495138 0.4230425 0.4720729
#> [197] 0.4254027 0.4344290 0.4199451 0.4194151 0.4594422 0.4136277 0.4243551
#> [204] 0.4326274 0.4387952 0.4120341 0.4488944 0.4342210 0.4308897 0.4179986
#> [211] 0.4193605 0.4222934 0.4590372 0.4034661 0.4598453 0.4586805 0.4212564
#> [218] 0.4469110 0.4326274 0.4165906 0.4365880 0.4053215 0.4245128 0.4198039
#> [225] 0.4211385 0.4340273 0.4430050 0.4193605 0.4455053 0.4329999 0.4130683
#> [232] 0.4118355 0.4280902 0.4295881 0.4277836 0.4196652 0.4238758 0.4165906
#> [239] 0.4262081 0.4277836 0.4555444 0.4334536 0.4295881 0.4180154 0.4330907
#> [246] 0.4034661 0.4179981 0.4165906 0.4136277 0.4467236 0.4230425 0.4262081
#> [253] 0.4585301 0.4357936 0.4323469 0.4555444 0.4102056 0.4279942 0.4588032
#> [260] 0.4326274 0.4588127 0.4243551 0.4771331 0.4053215 0.4346891 0.4521496
#> [267] 0.4585096 0.4179986 0.4112491 0.4421761 0.4326274 0.4118355 0.4280902
#> [274] 0.4165906 0.4395877 0.4699748 0.4730300 0.4279942 0.4335942 0.4751840
#> [281] 0.4179981 0.4555278 0.4326274 0.4585301 0.4585613 0.4180694 0.4102056
#> [288] 0.4120341 0.4179986 0.4277836 0.4342025 0.4198039 0.4334536 0.4118355
#> [295] 0.4245128 0.4165906 0.4136277 0.4318621 0.4382302 0.4809427 0.4165906
#> [302] 0.4334536 0.4585301 0.4724037 0.4277836 0.4566353 0.4485966 0.4194151
#> [309] 0.4500087 0.4513309 0.4598453 0.4118355 0.4613698 0.4272275 0.4165906
#> [316] 0.4260299 0.4177106 0.4187936 0.4120341 0.4177106 0.4190151 0.4198769
#> [323] 0.4042617 0.4042617 0.4112491 0.4112491 0.4034661 0.4044682 0.4181355
#> [330] 0.4190151 0.4042617 0.4112491 0.4044682 0.4044682 0.4199451 0.4187936
#> [337] 0.4034661 0.4112491 0.4187936 0.4245449 0.4042617 0.4198769 0.4130683
#> [344] 0.4123313 0.4044682 0.4123313 0.4199451 0.4042617 0.4130683 0.4112491
#> [351] 0.4190151 0.4044682 0.4199451 0.4044682 0.4123313 0.4034661 0.4157144
#> [358] 0.4034661 0.4199451 0.4214717 0.4330907 0.4326274 0.4222934 0.4454765
#> [365] 0.4629047 0.4555444 0.4198769 0.4034661 0.4034661 0.4172327 0.4751840
#> [372] 0.4586805 0.4180154 0.4179986 0.4179981 0.4112491 0.4277836 0.4469110
#> [379] 0.4042617 0.4190151 0.4044682 0.4613698 0.4357936 0.4585613 0.4330907
#> [386] 0.4079500 0.4044682 0.4340273 0.4259103 0.4044682 0.4262081 0.4295881
#> [393] 0.4469110 0.4444685 0.4295881 0.4157144 0.4327484 0.4336368 0.4044682
#> [400] 0.4120341 0.4277836 0.4521496 0.4588032 0.4102056 0.4726047 0.4199451
#> [407] 0.4123313 0.4034661 0.4455053 0.4255752 0.4342025 0.4382670 0.4326274
#> [414] 0.4199451 0.4260299 0.4198769 0.4402919 0.4340273 0.4357936 0.4199451
#> [421] 0.4330907 0.4585613 0.4323254 0.4586805 0.4277836 0.4469110 0.4172327
#> [428] 0.4699748 0.4220477 0.4262081 0.4211385 0.4238758 0.4537749 0.4259103
#> [435] 0.4254027 0.4336149 0.4165906 0.4330907 0.4102056 0.4488944 0.4157144
#> [442] 0.4330907 0.4259103 0.4245128 0.4102056