Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4450091 0.4245330 0.4643628 0.4406675 0.4369834 0.4245330 0.4615495
#>   [8] 0.4358914 0.4916720 0.4631434 0.4165905 0.4729655 0.4210937 0.4235861
#>  [15] 0.4220374 0.4066661 0.4841530 0.4161434 0.4809423 0.4663943 0.4331836
#>  [22] 0.5092988 0.4408453 0.4343943 0.4053048 0.4370808 0.4372103 0.4235861
#>  [29] 0.4066385 0.4343943 0.4066385 0.4856721 0.4372103 0.4530346 0.4653053
#>  [36] 0.5119935 0.4516027 0.4615495 0.4650487 0.4173558 0.4573872 0.4165905
#>  [43] 0.4225745 0.4210937 0.4248994 0.4372103 0.4225745 0.4166909 0.4516027
#>  [50] 0.4615495 0.4360383 0.4369714 0.4218171 0.4653053 0.4371641 0.4066385
#>  [57] 0.4397905 0.4819889 0.5119935 0.4314771 0.4225745 0.4367447 0.4220445
#>  [64] 0.4406675 0.4607320 0.4302769 0.4406675 0.4309803 0.4780491 0.4802386
#>  [71] 0.4339767 0.5041676 0.4267761 0.4653053 0.4954266 0.4372103 0.4598806
#>  [78] 0.4221243 0.4545155 0.4314771 0.4653053 0.4496422 0.4727110 0.4841530
#>  [85] 0.4395912 0.4461290 0.4318439 0.4385487 0.4573872 0.4357704 0.4369714
#>  [92] 0.4066385 0.4372103 0.4573872 0.4603075 0.4653053 0.4152156 0.4867358
#>  [99] 0.4199727 0.4312169 0.4173558 0.4213720 0.4372103 0.4218171 0.4232313
#> [106] 0.4551345 0.4841530 0.4166909 0.4371641 0.4213720 0.4182092 0.4166909
#> [113] 0.4530346 0.4218171 0.4684320 0.4573872 0.4684320 0.4530346 0.4066661
#> [120] 0.4615495 0.4906720 0.4053048 0.4165905 0.4841530 0.4218171 0.4339767
#> [127] 0.4598806 0.4161434 0.4615495 0.4248994 0.4166909 0.4366269 0.4385487
#> [134] 0.4248994 0.4343943 0.4312169 0.4318439 0.4053048 0.4644092 0.4385487
#> [141] 0.4275211 0.4653053 0.4173558 0.4650487 0.4640442 0.4232313 0.4223006
#> [148] 0.4406675 0.4385487 0.4082804 0.4198181 0.4626102 0.4223006 0.4615495
#> [155] 0.4552952 0.4810588 0.4152156 0.4573872 0.4235861 0.4277402 0.4245330
#> [162] 0.4612277 0.4173558 0.4203457 0.4644092 0.4066661 0.4232313 0.4312169
#> [169] 0.4144972 0.4221243 0.4066385 0.4663943 0.4182092 0.4066661 0.4213720
#> [176] 0.4366269 0.4598806 0.4361840 0.4598806 0.4203457 0.4631434 0.4913830
#> [183] 0.4330632 0.4394021 0.4361840 0.4161434 0.4663943 0.4245330 0.4152156
#> [190] 0.4615306 0.4152156 0.4331255 0.4358304 0.4567381 0.4234160 0.4831542
#> [197] 0.4302769 0.4386053 0.4232313 0.4238947 0.4626684 0.4173558 0.4301272
#> [204] 0.4372103 0.4431626 0.4165905 0.4551885 0.4395912 0.4357704 0.4204287
#> [211] 0.4196977 0.4269698 0.4631434 0.4053048 0.4663943 0.4640442 0.4220445
#> [218] 0.4530346 0.4372103 0.4199727 0.4381431 0.4082804 0.4301883 0.4235861
#> [225] 0.4248994 0.4385487 0.4472018 0.4196977 0.4511898 0.4362903 0.4182092
#> [232] 0.4161434 0.4299021 0.4343943 0.4331836 0.4232727 0.4290288 0.4199727
#> [239] 0.4314771 0.4331836 0.4626102 0.4370808 0.4343943 0.4245330 0.4366269
#> [246] 0.4053048 0.4213720 0.4199727 0.4173558 0.4528588 0.4234160 0.4314771
#> [253] 0.4653053 0.4406675 0.4367447 0.4626102 0.4144972 0.4322682 0.4644092
#> [260] 0.4372103 0.4643628 0.4301272 0.4780248 0.4082804 0.4358304 0.4598806
#> [267] 0.4628540 0.4204287 0.4152156 0.4453774 0.4372103 0.4161434 0.4299021
#> [274] 0.4199727 0.4442905 0.4749345 0.4784738 0.4322682 0.4358914 0.4841530
#> [281] 0.4213720 0.4603075 0.4372103 0.4653053 0.4640612 0.4194435 0.4144972
#> [288] 0.4165905 0.4204287 0.4331836 0.4361840 0.4235861 0.4370808 0.4161434
#> [295] 0.4301883 0.4199727 0.4173558 0.4360383 0.4425820 0.4904184 0.4199727
#> [302] 0.4370808 0.4653053 0.4799643 0.4331836 0.4572543 0.4532313 0.4238947
#> [309] 0.4564685 0.4552317 0.4663943 0.4161434 0.4684320 0.4307961 0.4199727
#> [316] 0.4318439 0.4210937 0.4225745 0.4165905 0.4210937 0.4218171 0.4230467
#> [323] 0.4066661 0.4066661 0.4152156 0.4152156 0.4053048 0.4066385 0.4205600
#> [330] 0.4218171 0.4066661 0.4152156 0.4066385 0.4066385 0.4232313 0.4225745
#> [337] 0.4053048 0.4152156 0.4225745 0.4279187 0.4066661 0.4230467 0.4182092
#> [344] 0.4166909 0.4066385 0.4166909 0.4232313 0.4066661 0.4182092 0.4152156
#> [351] 0.4218171 0.4066385 0.4232313 0.4066385 0.4166909 0.4053048 0.4221243
#> [358] 0.4053048 0.4232313 0.4275211 0.4366269 0.4372103 0.4269698 0.4484064
#> [365] 0.4661278 0.4626102 0.4230467 0.4053048 0.4053048 0.4198181 0.4841530
#> [372] 0.4640442 0.4245330 0.4204287 0.4213720 0.4152156 0.4331836 0.4530346
#> [379] 0.4066661 0.4218171 0.4066385 0.4684320 0.4406675 0.4640612 0.4366269
#> [386] 0.4121321 0.4066385 0.4385487 0.4308743 0.4066385 0.4314771 0.4343943
#> [393] 0.4530346 0.4496422 0.4343943 0.4221243 0.4369915 0.4369714 0.4066385
#> [400] 0.4165905 0.4331836 0.4598806 0.4644092 0.4144972 0.4802386 0.4232313
#> [407] 0.4166909 0.4053048 0.4511898 0.4295296 0.4361840 0.4394350 0.4372103
#> [414] 0.4232313 0.4318439 0.4230467 0.4441248 0.4385487 0.4406675 0.4232313
#> [421] 0.4366269 0.4640612 0.4371972 0.4640442 0.4331836 0.4530346 0.4198181
#> [428] 0.4749345 0.4277402 0.4314771 0.4248994 0.4290288 0.4615495 0.4308743
#> [435] 0.4302769 0.4371641 0.4199727 0.4366269 0.4144972 0.4551885 0.4221243
#> [442] 0.4366269 0.4308743 0.4301883 0.4144972