Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4422387 0.4192913 0.4588515 0.4364176 0.4349935 0.4192913 0.4534866
#>   [8] 0.4347283 0.4815534 0.4592505 0.4129214 0.4634866 0.4178354 0.4209506
#>  [15] 0.4192194 0.4055945 0.4743206 0.4131265 0.4733345 0.4590563 0.4284954
#>  [22] 0.4979455 0.4371148 0.4304273 0.4043260 0.4345170 0.4331611 0.4209506
#>  [29] 0.4053990 0.4304273 0.4053990 0.4783838 0.4331611 0.4474257 0.4579258
#>  [36] 0.5006543 0.4455792 0.4534866 0.4573698 0.4149748 0.4502117 0.4129214
#>  [43] 0.4189718 0.4178354 0.4227130 0.4331611 0.4189718 0.4128234 0.4455792
#>  [50] 0.4534866 0.4323234 0.4350566 0.4191806 0.4579258 0.4348359 0.4053990
#>  [57] 0.4354790 0.4734364 0.5006543 0.4270750 0.4189718 0.4333180 0.4218850
#>  [64] 0.4364176 0.4551688 0.4261184 0.4364176 0.4267067 0.4703924 0.4721764
#>  [71] 0.4286214 0.4927227 0.4235495 0.4579258 0.4849076 0.4331611 0.4520612
#>  [78] 0.4175294 0.4502464 0.4270750 0.4579258 0.4436040 0.4671917 0.4743206
#>  [85] 0.4347339 0.4438952 0.4266290 0.4344312 0.4502117 0.4312203 0.4350566
#>  [92] 0.4053990 0.4331611 0.4502117 0.4552978 0.4579258 0.4116787 0.4783936
#>  [99] 0.4176691 0.4280201 0.4149748 0.4189478 0.4331611 0.4191806 0.4205086
#> [106] 0.4487128 0.4743206 0.4128234 0.4348359 0.4189478 0.4149994 0.4128234
#> [113] 0.4474257 0.4191806 0.4608784 0.4502117 0.4608784 0.4474257 0.4055945
#> [120] 0.4534866 0.4816475 0.4043260 0.4129214 0.4743206 0.4191806 0.4286214
#> [127] 0.4520612 0.4131265 0.4534866 0.4227130 0.4128234 0.4336894 0.4344312
#> [134] 0.4227130 0.4304273 0.4280201 0.4266290 0.4043260 0.4587118 0.4344312
#> [141] 0.4224943 0.4579258 0.4149748 0.4573698 0.4589971 0.4205086 0.4189163
#> [148] 0.4364176 0.4344312 0.4076497 0.4183885 0.4553091 0.4189163 0.4534866
#> [155] 0.4486329 0.4713023 0.4116787 0.4502117 0.4209506 0.4234893 0.4192913
#> [162] 0.4562778 0.4149748 0.4192241 0.4587118 0.4055945 0.4205086 0.4280201
#> [169] 0.4117419 0.4175294 0.4053990 0.4590563 0.4149994 0.4055945 0.4189478
#> [176] 0.4336894 0.4520612 0.4355078 0.4520612 0.4192241 0.4592505 0.4807575
#> [183] 0.4301386 0.4364703 0.4355078 0.4131265 0.4590563 0.4192913 0.4116787
#> [190] 0.4563323 0.4116787 0.4284949 0.4353301 0.4501551 0.4235778 0.4730045
#> [197] 0.4261184 0.4348316 0.4205086 0.4209292 0.4590557 0.4149748 0.4252216
#> [204] 0.4331611 0.4387825 0.4129214 0.4484988 0.4347339 0.4312203 0.4198878
#> [211] 0.4203796 0.4236097 0.4592505 0.4043260 0.4590563 0.4589971 0.4218850
#> [218] 0.4474257 0.4331611 0.4176691 0.4368373 0.4076497 0.4251972 0.4209506
#> [225] 0.4227130 0.4344312 0.4433839 0.4203796 0.4451133 0.4341396 0.4149994
#> [232] 0.4131265 0.4289573 0.4304273 0.4284954 0.4211439 0.4252986 0.4176691
#> [239] 0.4270750 0.4284954 0.4553091 0.4345170 0.4304273 0.4192913 0.4336894
#> [246] 0.4043260 0.4189478 0.4176691 0.4149748 0.4468458 0.4235778 0.4270750
#> [253] 0.4579258 0.4364176 0.4333180 0.4553091 0.4117419 0.4285905 0.4587118
#> [260] 0.4331611 0.4588515 0.4252216 0.4773621 0.4076497 0.4353301 0.4520612
#> [267] 0.4585599 0.4198878 0.4116787 0.4424433 0.4331611 0.4131265 0.4289573
#> [274] 0.4176691 0.4405048 0.4687390 0.4722722 0.4285905 0.4347283 0.4743206
#> [281] 0.4189478 0.4552978 0.4331611 0.4579258 0.4579519 0.4196945 0.4117419
#> [288] 0.4129214 0.4198878 0.4284954 0.4355078 0.4209506 0.4345170 0.4131265
#> [295] 0.4251972 0.4176691 0.4149748 0.4323234 0.4380586 0.4816940 0.4176691
#> [302] 0.4345170 0.4579258 0.4714442 0.4284954 0.4565264 0.4495633 0.4209292
#> [309] 0.4497002 0.4506922 0.4590563 0.4131265 0.4608784 0.4287107 0.4176691
#> [316] 0.4266290 0.4178354 0.4189718 0.4129214 0.4178354 0.4191806 0.4201352
#> [323] 0.4055945 0.4055945 0.4116787 0.4116787 0.4043260 0.4053990 0.4180953
#> [330] 0.4191806 0.4055945 0.4116787 0.4053990 0.4053990 0.4205086 0.4189718
#> [337] 0.4043260 0.4116787 0.4189718 0.4247074 0.4055945 0.4201352 0.4149994
#> [344] 0.4128234 0.4053990 0.4128234 0.4205086 0.4055945 0.4149994 0.4116787
#> [351] 0.4191806 0.4053990 0.4205086 0.4053990 0.4128234 0.4043260 0.4175294
#> [358] 0.4043260 0.4205086 0.4224943 0.4336894 0.4331611 0.4236097 0.4455255
#> [365] 0.4618944 0.4553091 0.4201352 0.4043260 0.4043260 0.4183885 0.4743206
#> [372] 0.4589971 0.4192913 0.4198878 0.4189478 0.4116787 0.4284954 0.4474257
#> [379] 0.4055945 0.4191806 0.4053990 0.4608784 0.4364176 0.4579519 0.4336894
#> [386] 0.4100617 0.4053990 0.4344312 0.4270589 0.4053990 0.4270750 0.4304273
#> [393] 0.4474257 0.4436040 0.4304273 0.4175294 0.4339040 0.4350566 0.4053990
#> [400] 0.4129214 0.4284954 0.4520612 0.4587118 0.4117419 0.4721764 0.4205086
#> [407] 0.4128234 0.4043260 0.4451133 0.4266332 0.4355078 0.4384601 0.4331611
#> [414] 0.4205086 0.4266290 0.4201352 0.4406621 0.4344312 0.4364176 0.4205086
#> [421] 0.4336894 0.4579519 0.4325184 0.4589971 0.4284954 0.4474257 0.4183885
#> [428] 0.4687390 0.4234893 0.4270750 0.4227130 0.4252986 0.4534866 0.4270589
#> [435] 0.4261184 0.4348359 0.4176691 0.4336894 0.4117419 0.4484988 0.4175294
#> [442] 0.4336894 0.4270589 0.4251972 0.4117419