Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4376954 0.4155116 0.4625173 0.4332236 0.4313646 0.4155116 0.4588241
#>   [8] 0.4287479 0.4946645 0.4618931 0.4071438 0.4725309 0.4129660 0.4137039
#>  [15] 0.4141529 0.3951263 0.4851925 0.4050430 0.4812266 0.4653403 0.4243741
#>  [22] 0.5119269 0.4338493 0.4256320 0.3930689 0.4295955 0.4299466 0.4137039
#>  [29] 0.3956415 0.4256320 0.3956415 0.4843545 0.4299466 0.4498874 0.4640784
#>  [36] 0.5139752 0.4469775 0.4588241 0.4615085 0.4062866 0.4545163 0.4071438
#>  [43] 0.4156315 0.4129660 0.4161193 0.4299466 0.4156315 0.4076688 0.4469775
#>  [50] 0.4588241 0.4302200 0.4296801 0.4119832 0.4640784 0.4294727 0.3956415
#>  [57] 0.4320121 0.4790058 0.5139752 0.4229210 0.4156315 0.4308004 0.4128639
#>  [64] 0.4332236 0.4563456 0.4223438 0.4332236 0.4231933 0.4748777 0.4811336
#>  [71] 0.4260752 0.5067926 0.4185370 0.4640784 0.4973392 0.4299466 0.4572910
#>  [78] 0.4130687 0.4490918 0.4229210 0.4640784 0.4455107 0.4716761 0.4851925
#>  [85] 0.4338397 0.4394447 0.4247710 0.4313847 0.4545163 0.4301690 0.4296801
#>  [92] 0.3956415 0.4299466 0.4545163 0.4539140 0.4640784 0.4050227 0.4864130
#>  [99] 0.4103041 0.4234624 0.4062866 0.4117467 0.4299466 0.4119832 0.4141398
#> [106] 0.4492765 0.4851925 0.4076688 0.4294727 0.4117467 0.4089095 0.4076688
#> [113] 0.4498874 0.4119832 0.4669768 0.4545163 0.4669768 0.4498874 0.3951263
#> [120] 0.4588241 0.4916906 0.3930689 0.4071438 0.4851925 0.4119832 0.4260752
#> [127] 0.4572910 0.4050430 0.4588241 0.4161193 0.4076688 0.4291970 0.4313847
#> [134] 0.4161193 0.4256320 0.4234624 0.4247710 0.3930689 0.4628746 0.4313847
#> [141] 0.4183923 0.4640784 0.4062866 0.4615085 0.4625295 0.4141398 0.4150008
#> [148] 0.4332236 0.4313847 0.3969782 0.4098530 0.4599078 0.4150008 0.4588241
#> [155] 0.4491759 0.4823110 0.4050227 0.4545163 0.4137039 0.4208752 0.4155116
#> [162] 0.4573157 0.4062866 0.4102864 0.4628746 0.3951263 0.4141398 0.4234624
#> [169] 0.4034863 0.4130687 0.3956415 0.4653403 0.4089095 0.3951263 0.4117467
#> [176] 0.4291970 0.4572910 0.4292891 0.4572910 0.4102864 0.4618931 0.4926786
#> [183] 0.4260836 0.4341549 0.4292891 0.4050430 0.4653403 0.4155116 0.4050227
#> [190] 0.4556650 0.4050227 0.4260196 0.4287853 0.4512404 0.4147701 0.4827874
#> [197] 0.4223438 0.4320715 0.4141398 0.4167249 0.4612307 0.4062866 0.4231434
#> [204] 0.4299466 0.4362979 0.4071438 0.4486160 0.4338397 0.4301690 0.4106005
#> [211] 0.4100372 0.4192574 0.4618931 0.3930689 0.4653403 0.4625295 0.4128639
#> [218] 0.4498874 0.4299466 0.4103041 0.4315618 0.3969782 0.4224686 0.4137039
#> [225] 0.4161193 0.4313847 0.4419626 0.4100372 0.4477978 0.4305123 0.4089095
#> [232] 0.4050430 0.4207447 0.4256320 0.4243741 0.4151675 0.4203426 0.4103041
#> [239] 0.4229210 0.4243741 0.4599078 0.4295955 0.4256320 0.4155116 0.4291970
#> [246] 0.3930689 0.4117467 0.4103041 0.4062866 0.4484998 0.4147701 0.4229210
#> [253] 0.4640784 0.4332236 0.4308004 0.4599078 0.4034863 0.4243506 0.4628746
#> [260] 0.4299466 0.4625173 0.4231434 0.4759689 0.3969782 0.4287853 0.4572910
#> [267] 0.4614423 0.4106005 0.4050227 0.4366182 0.4299466 0.4050430 0.4207447
#> [274] 0.4103041 0.4359288 0.4740003 0.4791383 0.4243506 0.4287479 0.4851925
#> [281] 0.4117467 0.4539140 0.4299466 0.4640784 0.4626530 0.4097281 0.4034863
#> [288] 0.4071438 0.4106005 0.4243741 0.4292891 0.4137039 0.4295955 0.4050430
#> [295] 0.4224686 0.4103041 0.4062866 0.4302200 0.4372963 0.4912354 0.4103041
#> [302] 0.4295955 0.4640784 0.4809187 0.4243741 0.4519165 0.4485434 0.4167249
#> [309] 0.4505926 0.4497947 0.4653403 0.4050430 0.4669768 0.4233427 0.4103041
#> [316] 0.4247710 0.4129660 0.4156315 0.4071438 0.4129660 0.4119832 0.4144176
#> [323] 0.3951263 0.3951263 0.4050227 0.4050227 0.3930689 0.3956415 0.4114853
#> [330] 0.4119832 0.3951263 0.4050227 0.3956415 0.3956415 0.4141398 0.4156315
#> [337] 0.3930689 0.4050227 0.4156315 0.4205170 0.3951263 0.4144176 0.4089095
#> [344] 0.4076688 0.3956415 0.4076688 0.4141398 0.3951263 0.4089095 0.4050227
#> [351] 0.4119832 0.3956415 0.4141398 0.3956415 0.4076688 0.3930689 0.4130687
#> [358] 0.3930689 0.4141398 0.4183923 0.4291970 0.4299466 0.4192574 0.4418381
#> [365] 0.4659256 0.4599078 0.4144176 0.3930689 0.3930689 0.4098530 0.4851925
#> [372] 0.4625295 0.4155116 0.4106005 0.4117467 0.4050227 0.4243741 0.4498874
#> [379] 0.3951263 0.4119832 0.3956415 0.4669768 0.4332236 0.4626530 0.4291970
#> [386] 0.4010180 0.3956415 0.4313847 0.4228230 0.3956415 0.4229210 0.4256320
#> [393] 0.4498874 0.4455107 0.4256320 0.4130687 0.4308011 0.4296801 0.3956415
#> [400] 0.4071438 0.4243741 0.4572910 0.4628746 0.4034863 0.4811336 0.4141398
#> [407] 0.4076688 0.3930689 0.4477978 0.4221932 0.4292891 0.4334698 0.4299466
#> [414] 0.4141398 0.4247710 0.4144176 0.4360521 0.4313847 0.4332236 0.4141398
#> [421] 0.4291970 0.4626530 0.4316724 0.4625295 0.4243741 0.4498874 0.4098530
#> [428] 0.4740003 0.4208752 0.4229210 0.4161193 0.4203426 0.4588241 0.4228230
#> [435] 0.4223438 0.4294727 0.4103041 0.4291970 0.4034863 0.4486160 0.4130687
#> [442] 0.4291970 0.4228230 0.4224686 0.4034863
# }