Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4438474 0.4212340 0.4608067 0.4380005 0.4359331 0.4212340 0.4574545
#>   [8] 0.4348875 0.4851254 0.4607448 0.4147135 0.4681228 0.4199678 0.4218073
#>  [15] 0.4211145 0.4060405 0.4787557 0.4144240 0.4774497 0.4631946 0.4308698
#>  [22] 0.5030562 0.4397996 0.4323591 0.4051123 0.4354925 0.4354141 0.4218073
#>  [29] 0.4063950 0.4323591 0.4063950 0.4833609 0.4354141 0.4496003 0.4618571
#>  [36] 0.5053268 0.4489711 0.4574545 0.4618581 0.4158924 0.4536430 0.4147135
#>  [43] 0.4212568 0.4199678 0.4234127 0.4354141 0.4212568 0.4151306 0.4489711
#>  [50] 0.4574545 0.4341571 0.4352570 0.4209598 0.4618571 0.4352360 0.4063950
#>  [57] 0.4381035 0.4774440 0.5053268 0.4289518 0.4212568 0.4351218 0.4218738
#>  [64] 0.4380005 0.4577024 0.4279766 0.4380005 0.4289080 0.4749674 0.4759305
#>  [71] 0.4309227 0.4989382 0.4256554 0.4618571 0.4882871 0.4354141 0.4554968
#>  [78] 0.4195065 0.4526824 0.4289518 0.4618571 0.4468053 0.4704876 0.4787557
#>  [85] 0.4369592 0.4455828 0.4291139 0.4368426 0.4536430 0.4338323 0.4352570
#>  [92] 0.4063950 0.4354141 0.4536430 0.4582037 0.4618571 0.4137980 0.4819694
#>  [99] 0.4189142 0.4289657 0.4158924 0.4203842 0.4354141 0.4209598 0.4219161
#> [106] 0.4525556 0.4787557 0.4151306 0.4352360 0.4203842 0.4158557 0.4151306
#> [113] 0.4496003 0.4209598 0.4642009 0.4536430 0.4642009 0.4496003 0.4060405
#> [120] 0.4574545 0.4851508 0.4051123 0.4147135 0.4787557 0.4209598 0.4309227
#> [127] 0.4554968 0.4144240 0.4574545 0.4234127 0.4151306 0.4347364 0.4368426
#> [134] 0.4234127 0.4323591 0.4289657 0.4291139 0.4051123 0.4612546 0.4368426
#> [141] 0.4247278 0.4618571 0.4158924 0.4618581 0.4607376 0.4219161 0.4208450
#> [148] 0.4380005 0.4368426 0.4072671 0.4187778 0.4588491 0.4208450 0.4574545
#> [155] 0.4517389 0.4763857 0.4137980 0.4536430 0.4218073 0.4254399 0.4212340
#> [162] 0.4580091 0.4158924 0.4196110 0.4612546 0.4060405 0.4219161 0.4289657
#> [169] 0.4125571 0.4195065 0.4063950 0.4631946 0.4158557 0.4060405 0.4203842
#> [176] 0.4347364 0.4554968 0.4355032 0.4554968 0.4196110 0.4607448 0.4853692
#> [183] 0.4315223 0.4379983 0.4355032 0.4144240 0.4631946 0.4212340 0.4137980
#> [190] 0.4590209 0.4137980 0.4307120 0.4353701 0.4525120 0.4237064 0.4771158
#> [197] 0.4279766 0.4367936 0.4219161 0.4217875 0.4604908 0.4158924 0.4271824
#> [204] 0.4354141 0.4412159 0.4147135 0.4525476 0.4369592 0.4338323 0.4195342
#> [211] 0.4198836 0.4253350 0.4607448 0.4051123 0.4631946 0.4607376 0.4218738
#> [218] 0.4496003 0.4354141 0.4189142 0.4373486 0.4072671 0.4278178 0.4218073
#> [225] 0.4234127 0.4368426 0.4462927 0.4198836 0.4479517 0.4350115 0.4158557
#> [232] 0.4144240 0.4300922 0.4323591 0.4308698 0.4216875 0.4271297 0.4189142
#> [239] 0.4289518 0.4308698 0.4588491 0.4354925 0.4323591 0.4212340 0.4347364
#> [246] 0.4051123 0.4203842 0.4189142 0.4158924 0.4504809 0.4237064 0.4289518
#> [253] 0.4618571 0.4380005 0.4351218 0.4588491 0.4125571 0.4304733 0.4612546
#> [260] 0.4354141 0.4608067 0.4271824 0.4782973 0.4072671 0.4353701 0.4554968
#> [267] 0.4602975 0.4195342 0.4137980 0.4449967 0.4354141 0.4144240 0.4300922
#> [274] 0.4189142 0.4421152 0.4718070 0.4752441 0.4304733 0.4348875 0.4787557
#> [281] 0.4203842 0.4582037 0.4354141 0.4618571 0.4606466 0.4191895 0.4125571
#> [288] 0.4147135 0.4195342 0.4308698 0.4355032 0.4218073 0.4354925 0.4144240
#> [295] 0.4278178 0.4189142 0.4158924 0.4341571 0.4408892 0.4846881 0.4189142
#> [302] 0.4354925 0.4618571 0.4752788 0.4308698 0.4573048 0.4516103 0.4217875
#> [309] 0.4535264 0.4536031 0.4631946 0.4144240 0.4642009 0.4295041 0.4189142
#> [316] 0.4291139 0.4199678 0.4212568 0.4147135 0.4199678 0.4209598 0.4220499
#> [323] 0.4060405 0.4060405 0.4137980 0.4137980 0.4051123 0.4063950 0.4198518
#> [330] 0.4209598 0.4060405 0.4137980 0.4063950 0.4063950 0.4219161 0.4212568
#> [337] 0.4051123 0.4137980 0.4212568 0.4264771 0.4060405 0.4220499 0.4158557
#> [344] 0.4151306 0.4063950 0.4151306 0.4219161 0.4060405 0.4158557 0.4137980
#> [351] 0.4209598 0.4063950 0.4219161 0.4063950 0.4151306 0.4051123 0.4195065
#> [358] 0.4051123 0.4219161 0.4247278 0.4347364 0.4354141 0.4253350 0.4469729
#> [365] 0.4641291 0.4588491 0.4220499 0.4051123 0.4051123 0.4187778 0.4787557
#> [372] 0.4607376 0.4212340 0.4195342 0.4203842 0.4137980 0.4308698 0.4496003
#> [379] 0.4060405 0.4209598 0.4063950 0.4642009 0.4380005 0.4606466 0.4347364
#> [386] 0.4108665 0.4063950 0.4368426 0.4288862 0.4063950 0.4289518 0.4323591
#> [393] 0.4496003 0.4468053 0.4323591 0.4195065 0.4350510 0.4352570 0.4063950
#> [400] 0.4147135 0.4308698 0.4554968 0.4612546 0.4125571 0.4759305 0.4219161
#> [407] 0.4151306 0.4051123 0.4479517 0.4274455 0.4355032 0.4391651 0.4354141
#> [414] 0.4219161 0.4291139 0.4220499 0.4430843 0.4368426 0.4380005 0.4219161
#> [421] 0.4347364 0.4606466 0.4352861 0.4607376 0.4308698 0.4496003 0.4187778
#> [428] 0.4718070 0.4254399 0.4289518 0.4234127 0.4271297 0.4574545 0.4288862
#> [435] 0.4279766 0.4352360 0.4189142 0.4347364 0.4125571 0.4525476 0.4195065
#> [442] 0.4347364 0.4288862 0.4278178 0.4125571
# }