Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
plotBart:::propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4407642 0.4169448 0.4588082 0.4345991 0.4322096 0.4169448 0.4539938
#>   [8] 0.4324075 0.4811622 0.4591419 0.4095892 0.4636441 0.4147515 0.4180793
#>  [15] 0.4170114 0.4017380 0.4745108 0.4105658 0.4756097 0.4595201 0.4270534
#>  [22] 0.5000643 0.4358269 0.4286918 0.4006640 0.4326287 0.4315498 0.4180793
#>  [29] 0.4018695 0.4286918 0.4018695 0.4777660 0.4315498 0.4462576 0.4583742
#>  [36] 0.5025094 0.4434293 0.4539938 0.4582937 0.4121149 0.4503867 0.4095892
#>  [43] 0.4160609 0.4147515 0.4196062 0.4315498 0.4160609 0.4097990 0.4434293
#>  [50] 0.4539938 0.4302854 0.4327933 0.4164761 0.4583742 0.4325660 0.4018695
#>  [57] 0.4336326 0.4722880 0.5025094 0.4254450 0.4160609 0.4309137 0.4192005
#>  [64] 0.4345991 0.4520891 0.4241856 0.4345991 0.4252285 0.4713224 0.4721545
#>  [71] 0.4262227 0.4963119 0.4206378 0.4583742 0.4851352 0.4315498 0.4523512
#>  [78] 0.4150086 0.4500817 0.4254450 0.4583742 0.4428841 0.4686228 0.4745108
#>  [85] 0.4321507 0.4429977 0.4244249 0.4327627 0.4503867 0.4289311 0.4327933
#>  [92] 0.4018695 0.4315498 0.4503867 0.4565356 0.4583742 0.4085340 0.4809227
#>  [99] 0.4149393 0.4251782 0.4121149 0.4161978 0.4315498 0.4164761 0.4178130
#> [106] 0.4491649 0.4745108 0.4097990 0.4325660 0.4161978 0.4112497 0.4097990
#> [113] 0.4462576 0.4164761 0.4612077 0.4503867 0.4612077 0.4462576 0.4017380
#> [120] 0.4539938 0.4813152 0.4006640 0.4095892 0.4745108 0.4164761 0.4262227
#> [127] 0.4523512 0.4105658 0.4539938 0.4196062 0.4097990 0.4323062 0.4327627
#> [134] 0.4196062 0.4286918 0.4251782 0.4244249 0.4006640 0.4589280 0.4327627
#> [141] 0.4200867 0.4583742 0.4121149 0.4582937 0.4589279 0.4178130 0.4156893
#> [148] 0.4345991 0.4327627 0.4034491 0.4159476 0.4555322 0.4156893 0.4539938
#> [155] 0.4484519 0.4716084 0.4085340 0.4503867 0.4180793 0.4209354 0.4169448
#> [162] 0.4527173 0.4121149 0.4164046 0.4589280 0.4017380 0.4178130 0.4251782
#> [169] 0.4090238 0.4150086 0.4018695 0.4595201 0.4112497 0.4017380 0.4161978
#> [176] 0.4323062 0.4523512 0.4332070 0.4523512 0.4164046 0.4591419 0.4805898
#> [183] 0.4278663 0.4343320 0.4332070 0.4105658 0.4595201 0.4169448 0.4085340
#> [190] 0.4577753 0.4085340 0.4259499 0.4336097 0.4470079 0.4209105 0.4726627
#> [197] 0.4241856 0.4329095 0.4178130 0.4172823 0.4593630 0.4121149 0.4228239
#> [204] 0.4315498 0.4379076 0.4095892 0.4494728 0.4321507 0.4289311 0.4166776
#> [211] 0.4176476 0.4218207 0.4591419 0.4006640 0.4595201 0.4589279 0.4192005
#> [218] 0.4462576 0.4315498 0.4149393 0.4351343 0.4034491 0.4229617 0.4180793
#> [225] 0.4196062 0.4327627 0.4422331 0.4176476 0.4443891 0.4312396 0.4112497
#> [232] 0.4105658 0.4271879 0.4286918 0.4270534 0.4181531 0.4235357 0.4149393
#> [239] 0.4254450 0.4270534 0.4555322 0.4326287 0.4286918 0.4169448 0.4323062
#> [246] 0.4006640 0.4161978 0.4149393 0.4121149 0.4446152 0.4209105 0.4254450
#> [253] 0.4583742 0.4345991 0.4309137 0.4555322 0.4090238 0.4268002 0.4589280
#> [260] 0.4315498 0.4588082 0.4228239 0.4747802 0.4034491 0.4336097 0.4523512
#> [267] 0.4584930 0.4166776 0.4085340 0.4424225 0.4315498 0.4105658 0.4271879
#> [274] 0.4149393 0.4399110 0.4705254 0.4723823 0.4268002 0.4324075 0.4745108
#> [281] 0.4161978 0.4565356 0.4315498 0.4583742 0.4587026 0.4163537 0.4090238
#> [288] 0.4095892 0.4166776 0.4270534 0.4332070 0.4180793 0.4326287 0.4105658
#> [295] 0.4229617 0.4149393 0.4121149 0.4302854 0.4360327 0.4812093 0.4149393
#> [302] 0.4326287 0.4583742 0.4719150 0.4270534 0.4565515 0.4468465 0.4172823
#> [309] 0.4508111 0.4499848 0.4595201 0.4105658 0.4612077 0.4257860 0.4149393
#> [316] 0.4244249 0.4147515 0.4160609 0.4095892 0.4147515 0.4164761 0.4175620
#> [323] 0.4017380 0.4017380 0.4085340 0.4085340 0.4006640 0.4018695 0.4156441
#> [330] 0.4164761 0.4017380 0.4085340 0.4018695 0.4018695 0.4178130 0.4160609
#> [337] 0.4006640 0.4085340 0.4160609 0.4215679 0.4017380 0.4175620 0.4112497
#> [344] 0.4097990 0.4018695 0.4097990 0.4178130 0.4017380 0.4112497 0.4085340
#> [351] 0.4164761 0.4018695 0.4178130 0.4018695 0.4097990 0.4006640 0.4150086
#> [358] 0.4006640 0.4178130 0.4200867 0.4323062 0.4315498 0.4218207 0.4449927
#> [365] 0.4625262 0.4555322 0.4175620 0.4006640 0.4006640 0.4159476 0.4745108
#> [372] 0.4589279 0.4169448 0.4166776 0.4161978 0.4085340 0.4270534 0.4462576
#> [379] 0.4017380 0.4164761 0.4018695 0.4612077 0.4345991 0.4587026 0.4323062
#> [386] 0.4071880 0.4018695 0.4327627 0.4247009 0.4018695 0.4254450 0.4286918
#> [393] 0.4462576 0.4428841 0.4286918 0.4150086 0.4311100 0.4327933 0.4018695
#> [400] 0.4095892 0.4270534 0.4523512 0.4589280 0.4090238 0.4721545 0.4178130
#> [407] 0.4097990 0.4006640 0.4443891 0.4231508 0.4332070 0.4368293 0.4315498
#> [414] 0.4178130 0.4244249 0.4175620 0.4406031 0.4327627 0.4345991 0.4178130
#> [421] 0.4323062 0.4587026 0.4302323 0.4589279 0.4270534 0.4462576 0.4159476
#> [428] 0.4705254 0.4209354 0.4254450 0.4196062 0.4235357 0.4539938 0.4247009
#> [435] 0.4241856 0.4325660 0.4149393 0.4323062 0.4090238 0.4494728 0.4150086
#> [442] 0.4323062 0.4247009 0.4229617 0.4090238
# }