Calculate propensity scores using BART

Calculates propensity scores using Bayesian Additive Regression Trees via bartCause::bartc.

Usage

propensity_scores(.data, treatment, response, confounders)

Arguments

.data

dataframe

treatment

character. Name of the treatment column within .data

response

character. Name of the response column within .data

confounders

character list of column names denoting confounders within .data

Value

a numeric vector of propensity scores

See also

Examples

# \dontrun{
data(lalonde)
plotBart:::propensity_scores(
  .data = lalonde,
  treatment = 'treat',
  response = 're78',
  confounders = c('age', 'educ')
)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#>   [1] 0.4601454 0.4457275 0.4691688 0.4554012 0.4551812 0.4457275 0.4658208
#>   [8] 0.4548903 0.4839160 0.4696233 0.4420771 0.4720648 0.4446631 0.4461334
#>  [15] 0.4459429 0.4368825 0.4788174 0.4417377 0.4794018 0.4687359 0.4509667
#>  [22] 0.4946796 0.4578830 0.4517816 0.4359752 0.4546902 0.4533840 0.4461334
#>  [29] 0.4367946 0.4517816 0.4367946 0.4843539 0.4533840 0.4614984 0.4680783
#>  [36] 0.4960029 0.4623891 0.4658208 0.4705566 0.4425425 0.4631231 0.4420771
#>  [43] 0.4455518 0.4446631 0.4467276 0.4533840 0.4455518 0.4420351 0.4623891
#>  [50] 0.4658208 0.4538342 0.4546926 0.4448521 0.4680783 0.4548420 0.4367946
#>  [57] 0.4567304 0.4809422 0.4960029 0.4497809 0.4455518 0.4543313 0.4473981
#>  [64] 0.4554012 0.4692608 0.4504630 0.4554012 0.4506476 0.4789633 0.4779592
#>  [71] 0.4514646 0.4921511 0.4491451 0.4680783 0.4861420 0.4533840 0.4646425
#>  [78] 0.4443740 0.4653813 0.4497809 0.4680783 0.4595127 0.4753835 0.4788174
#>  [85] 0.4547875 0.4612906 0.4502154 0.4541769 0.4631231 0.4525561 0.4546926
#>  [92] 0.4367946 0.4533840 0.4631231 0.4684195 0.4680783 0.4411400 0.4819243
#>  [99] 0.4440810 0.4510103 0.4425425 0.4449395 0.4533840 0.4448521 0.4458457
#> [106] 0.4649094 0.4788174 0.4420351 0.4548420 0.4449395 0.4428648 0.4420351
#> [113] 0.4614984 0.4448521 0.4698828 0.4631231 0.4698828 0.4614984 0.4368825
#> [120] 0.4658208 0.4835162 0.4359752 0.4420771 0.4788174 0.4448521 0.4514646
#> [127] 0.4646425 0.4417377 0.4658208 0.4467276 0.4420351 0.4542679 0.4541769
#> [134] 0.4467276 0.4517816 0.4510103 0.4502154 0.4359752 0.4690226 0.4541769
#> [141] 0.4476700 0.4680783 0.4425425 0.4705566 0.4689697 0.4458457 0.4454619
#> [148] 0.4554012 0.4541769 0.4376845 0.4450830 0.4664526 0.4454619 0.4658208
#> [155] 0.4644879 0.4770581 0.4411400 0.4631231 0.4461334 0.4476757 0.4457275
#> [162] 0.4693928 0.4425425 0.4455776 0.4690226 0.4368825 0.4458457 0.4510103
#> [169] 0.4405743 0.4443740 0.4367946 0.4687359 0.4428648 0.4368825 0.4449395
#> [176] 0.4542679 0.4646425 0.4554162 0.4646425 0.4455776 0.4696233 0.4825519
#> [183] 0.4529689 0.4563918 0.4554162 0.4417377 0.4687359 0.4457275 0.4411400
#> [190] 0.4689894 0.4411400 0.4509373 0.4555059 0.4655228 0.4489313 0.4775189
#> [197] 0.4504630 0.4551377 0.4458457 0.4461810 0.4696056 0.4425425 0.4490504
#> [204] 0.4533840 0.4590548 0.4420771 0.4650810 0.4547875 0.4525561 0.4455583
#> [211] 0.4464564 0.4480850 0.4696233 0.4359752 0.4687359 0.4689697 0.4473981
#> [218] 0.4614984 0.4533840 0.4440810 0.4564148 0.4376845 0.4492478 0.4461334
#> [225] 0.4467276 0.4541769 0.4613661 0.4464564 0.4605563 0.4546441 0.4428648
#> [232] 0.4417377 0.4522725 0.4517816 0.4509667 0.4465655 0.4483224 0.4440810
#> [239] 0.4497809 0.4509667 0.4664526 0.4546902 0.4517816 0.4457275 0.4542679
#> [246] 0.4359752 0.4449395 0.4440810 0.4425425 0.4632645 0.4489313 0.4497809
#> [253] 0.4680783 0.4554012 0.4543313 0.4664526 0.4405743 0.4513413 0.4690226
#> [260] 0.4533840 0.4691688 0.4490504 0.4838141 0.4376845 0.4555059 0.4646425
#> [267] 0.4691632 0.4455583 0.4411400 0.4610348 0.4533840 0.4417377 0.4522725
#> [274] 0.4440810 0.4589392 0.4765264 0.4784662 0.4513413 0.4548903 0.4788174
#> [281] 0.4449395 0.4684195 0.4533840 0.4680783 0.4686551 0.4457416 0.4405743
#> [288] 0.4420771 0.4455583 0.4509667 0.4554162 0.4461334 0.4546902 0.4417377
#> [295] 0.4492478 0.4440810 0.4425425 0.4538342 0.4578223 0.4836901 0.4440810
#> [302] 0.4546902 0.4680783 0.4776147 0.4509667 0.4695756 0.4652143 0.4461810
#> [309] 0.4656625 0.4672099 0.4687359 0.4417377 0.4698828 0.4518098 0.4440810
#> [316] 0.4502154 0.4446631 0.4455518 0.4420771 0.4446631 0.4448521 0.4455958
#> [323] 0.4368825 0.4368825 0.4411400 0.4411400 0.4359752 0.4367946 0.4451055
#> [330] 0.4448521 0.4368825 0.4411400 0.4367946 0.4367946 0.4458457 0.4455518
#> [337] 0.4359752 0.4411400 0.4455518 0.4498130 0.4368825 0.4455958 0.4428648
#> [344] 0.4420351 0.4367946 0.4420351 0.4458457 0.4368825 0.4428648 0.4411400
#> [351] 0.4448521 0.4367946 0.4458457 0.4367946 0.4420351 0.4359752 0.4443740
#> [358] 0.4359752 0.4458457 0.4476700 0.4542679 0.4533840 0.4480850 0.4624591
#> [365] 0.4719488 0.4664526 0.4455958 0.4359752 0.4359752 0.4450830 0.4788174
#> [372] 0.4689697 0.4457275 0.4455583 0.4449395 0.4411400 0.4509667 0.4614984
#> [379] 0.4368825 0.4448521 0.4367946 0.4698828 0.4554012 0.4686551 0.4542679
#> [386] 0.4392433 0.4367946 0.4541769 0.4509483 0.4367946 0.4497809 0.4517816
#> [393] 0.4614984 0.4595127 0.4517816 0.4443740 0.4545065 0.4546926 0.4367946
#> [400] 0.4420771 0.4509667 0.4646425 0.4690226 0.4405743 0.4779592 0.4458457
#> [407] 0.4420351 0.4359752 0.4605563 0.4505514 0.4554162 0.4579215 0.4533840
#> [414] 0.4458457 0.4502154 0.4455958 0.4593505 0.4541769 0.4554012 0.4458457
#> [421] 0.4542679 0.4686551 0.4534500 0.4689697 0.4509667 0.4614984 0.4450830
#> [428] 0.4765264 0.4476757 0.4497809 0.4467276 0.4483224 0.4658208 0.4509483
#> [435] 0.4504630 0.4548420 0.4440810 0.4542679 0.4405743 0.4650810 0.4443740
#> [442] 0.4542679 0.4509483 0.4492478 0.4405743
# }